% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{alphahat}
\alias{alphahat}
\title{Extract alphahat from a KFS object}
\usage{
alphahat(object)
}
\arguments{
\item{object}{A \code{KFS} object.}
}
\value{
The smoothed state estimates \code{alphahat}.
}
\description{
Accessor method to access \code{alphahat}, the smoothed state
estimates, from a fitted \code{KFS} object.
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
conv <- xts_to_idx(gauteng)
model <- SSModelDynamicGompertz$new(Y = conv$series, q = 0.005,
                                    calendar = conv$calendar,
                                    end = idx_to_pos(conv$calendar, "2020-07-20"))
res <- estimate(model)
alphahat(output(res))

}
