% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_series.R
\name{as.matrix.idx_series}
\alias{as.matrix.idx_series}
\title{Coerce an \code{idx_series} to a matrix}
\usage{
\method{as.matrix}{idx_series}(x, ...)
}
\arguments{
\item{x}{An \code{idx_series} object.}

\item{...}{Unused.}
}
\description{
Coerce an \code{idx_series} to a matrix
}
