% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{att}
\alias{att}
\title{Extract filtered state estimates from a KFS object}
\usage{
att(object)
}
\arguments{
\item{object}{A \code{KFS} object.}
}
\value{
The filtered state estimates.
}
\description{
Accessor method to access the filtered state estimates in a
\code{KFS} object.
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
conv <- xts_to_idx(gauteng)
model <- SSModelDynamicGompertz$new(Y = conv$series, q = 0.005,
                                    calendar = conv$calendar,
                                    end = idx_to_pos(conv$calendar, "2020-07-20"))
res <- estimate(model)
att(output(res))
}
