% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cross_val}
\alias{cross_val}
\title{Walk-Forward Validation for Model Comparison Using Mean Absolute
Percentage Error (MAPE)}
\usage{
cross_val(
  Y,
  model_list,
  est.end,
  n.ahead = 7,
  n.estimate = 1,
  gap = 1,
  xpred_targ.full = NULL,
  xpred_lead.full = NULL,
  LeadIndCol = 1,
  criterion = "mape"
)
}
\arguments{
\item{Y}{An \code{idx_series} representing the cumulative data series. If
a Leading Indicator model is compared, Y should include columns for both
the leading indicator and the target variable. The specific column for
the leading indicator can be designated using the \code{LeadIndCol}
parameter.}

\item{model_list}{A list containing \code{SSModelDynamicGompertz} or
\code{SSModelLeadingIndicator} objects, to be compared in a cross
validation procedure.}

\item{est.end}{The initial estimation end position for model fitting.
Starting from this position, the function re-estimates the model and
evaluates the performance for each lag in \code{all_lags} every
\code{gap} positions, over \code{n.estimate} steps.}

\item{n.ahead}{Integer specifying the number of positions to forecast
ahead for MAPE evaluation.}

\item{n.estimate}{Integer indicating the total number of walk-forward
validation steps to report.}

\item{gap}{Integer specifying the position gap between two successive
validations, where the model is re-estimated and evaluated during the
walk-forward validation.}

\item{xpred_targ.full}{An \code{idx_series} containing the values of
exogenous variables for the target variable over the estimation and
prediction time frame.}

\item{xpred_lead.full}{(Only required for leading indicator models) An
\code{idx_series} containing the values of exogenous variables for the
leading indicator over the estimation and prediction time frame.}

\item{LeadIndCol}{(Only required for leading indicator models) Integer
representing the column number in \code{Y} that contains the leading
indicator.}

\item{criterion}{A string object indicating how to compare between
different models. Available choices are "mape" (by default), "smape",
"mae" and "rmse".}
}
\value{
A table summarizing the chosen error metric for each model in
\code{model_list} across the specified positions.
}
\description{
This function performs a walk-forward validation to compare
forecasting performance across different models specified by the user.
It returns a data frame of a user-specified error metric (e.g. MAPE, MAE)
for forecasts \code{n.ahead} positions ahead, using the given models with
varying end positions.
}
