% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df2ldl}
\alias{df2ldl}
\title{Compute log growth rate of cumulated dataset}
\usage{
df2ldl(dt)
}
\arguments{
\item{dt}{Cumulated data series, as an \code{idx_series} with exactly one
column.}
}
\value{
An \code{idx_series} of log growth rates of the cumulated
variable inputted via the parameter \code{dt}.
}
\description{
Helper method to compute the log growth rates of cumulated
variables.
}
\examples{
x <- idx_series(cumsum(rpois(30, 5)) + 1)
df2ldl(x)

}
