% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df2ldl_lead}
\alias{df2ldl_lead}
\title{Compute successive increments and log growth rate of 2-variable
cumulated dataset}
\usage{
df2ldl_lead(data, LeadIndCol = 1)
}
\arguments{
\item{data}{Cumulated data series as an \code{idx_series} with 2 columns:
leading indicator and target variable. Can specify which column is
leading indicator by \code{LeadIndCol} parameter.}

\item{LeadIndCol}{Column number of \code{data} that contains the leading
indicator. An integer that can only take values 1 (by default) or 2.}
}
\value{
A list of \code{idx_series} with the original cumulative
variables, successive increments and log growth rates: \code{cLead},
\code{cTarg}, \code{newLead}, \code{newTarg}, \code{LDLlead},
\code{LDLtarg}.
}
\description{
Helper method to compute the successive increments and log
growth rates of cumulated variables. It will compute the successive
increments and log cumulative growth rate for each column in the
2-column series, which will then be used to predict or estimate with the
leading indicator model.
}
