% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_weather_2021}
\alias{england_weather_2021}
\title{Weather data of England.}
\format{
An object of class \code{"xts"};
\describe{
\item{TempC}{Mean air temperature in degree Celsius at 2 meters}
\item{TempC_max}{Maximum Temperature in degree Celsius at 2 meters}
\item{RelHumid}{Relative humidity (\%)}
\item{WindSpeed}{Wind speed in m/s at 10 meters }
}
}
\usage{
data(england_weather_2021)
}
\description{
Weather data of England.
}
\examples{
data(england_weather_2021)
# plot daily cases
plot(diff(gauteng_weather_2021))
}
\references{
Downloaded from https://power.larc.nasa.gov/data-access-viewer/
}
\keyword{datasets}
