% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{estimate}
\alias{estimate}
\title{Estimate an SSModelDynamicGompertz or SSModelLeadingIndicator model}
\usage{
estimate(model)
}
\arguments{
\item{model}{A \code{SSModelDynamicGompertz} or \code{SSModelLeadingIndicator}
object.}
}
\value{
A \code{FilterResults} or \code{FilterResultsLI} object containing the
estimation results.
}
\description{
Accessor method that calls the \code{estimate} method of a
\code{SSModelDynamicGompertz} or \code{SSModelLeadingIndicator} object.
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
model <- SSModelDynamicGompertz$new(Y = gauteng, q = 0.005,
                                    end.date = as.Date("2020-07-20"))
estimate(model)

}
