% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gauteng}
\alias{gauteng}
\title{Cumlative cases of Covid-19 in the South African provence of Gauteng.}
\format{
An object of class \code{"xts"};
\describe{
\item{Cases}{Cumulative cases of Covid-19 from 10th March 2020}
}
}
\usage{
data(gauteng)
}
\description{
Cumlative cases of Covid-19 in the South African provence of Gauteng.
}
\examples{
data(gauteng)
# plot daily cases
plot(diff(gauteng))
}
\references{
tba
}
\keyword{datasets}
