% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gauteng_weather}
\alias{gauteng_weather}
\title{Weather data of the South African province of Gauteng.}
\format{
An object of class \code{"xts"};
\describe{
\item{TempC}{Mean air temperature in degree Celsius at 2 meters}
\item{TempC_max}{Maximum Temperature in degree Celsius at 2 meters}
\item{RelHumid}{Relative humidity (\%)}
\item{WindSpeed}{Wind speed in m/s at 10 meters }
}
}
\usage{
data(gauteng_weather)
}
\description{
Weather data of the South African province of Gauteng.
}
\examples{
data(gauteng_weather)
# plot daily cases
plot(diff(gauteng_weather))
}
\references{
Downloaded from https://power.larc.nasa.gov/data-access-viewer/
}
\keyword{datasets}
