% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_time_resolution}
\alias{get_time_resolution}
\title{Identify time resolution of given dates}
\usage{
get_time_resolution(dates)
}
\arguments{
\item{dates}{}
}
\description{
This function identifies the time resolution of the xts object
from its dates.
}
