% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_timeframe}
\alias{get_timeframe}
\title{Subsetting \code{idx_series} objects given start and end positions}
\usage{
get_timeframe(df, start, end = NULL)
}
\arguments{
\item{df}{An \code{idx_series} object, or \code{NULL} if no data are
supplied.}

\item{start}{Start position (integer) of the range.}

\item{end}{End position (integer) of the range. Defaults to the last
position in \code{df}.}
}
\value{
An \code{idx_series} containing the selected observations, or
\code{NULL} if \code{df} is \code{NULL}.
}
\description{
Helper method to subset an \code{idx_series} for a
specified range of integer positions.
}
\examples{
x <- idx_series(cumsum(rpois(30, 5)) + 1, start = 1)
get_timeframe(x, 5, 10)
get_timeframe(x, 5)

}
