% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_series.R
\name{idx_ncol}
\alias{idx_ncol}
\title{Number of columns in an \code{idx_series}}
\usage{
idx_ncol(x)
}
\arguments{
\item{x}{An \code{idx_series} object.}
}
\description{
\code{base::NCOL()} is not an S3 generic, so it will not
dispatch on \code{idx_series} objects. Use this function instead when
you need the number of columns/variables in an \code{idx_series}.
}
