% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_series.R
\name{idx_positions}
\alias{idx_positions}
\title{Integer positions of an \code{idx_series}}
\usage{
idx_positions(x)
}
\arguments{
\item{x}{An \code{idx_series} object.}
}
\value{
An integer vector.
}
\description{
Returns the full vector of integer positions associated
with each observation in \code{x}. This is the index-based analogue of
\code{zoo::index()}.
}
