% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_series.R
\name{idx_range}
\alias{idx_range}
\title{Integer position range spanned by an \code{idx_series}}
\usage{
idx_range(x)
}
\arguments{
\item{x}{An \code{idx_series} object.}
}
\value{
An integer vector of length 2: \code{c(first, last)}.
}
\description{
Returns the first and last integer position covered by
\code{x}, i.e. \code{c(x$start, x$start + length(x) - 1)}.
}
