% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_series.R
\name{idx_rbind}
\alias{idx_rbind}
\title{Bind two \code{idx_series} sequentially}
\usage{
idx_rbind(x, y)
}
\arguments{
\item{x, y}{\code{idx_series} objects to concatenate.}
}
\value{
A single, combined \code{idx_series}.
}
\description{
Concatenates \code{x} then \code{y} in position order.
\code{y} must start exactly one position after \code{x} ends.
}
