% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_calendar.R
\name{is_idx_calendar}
\alias{is_idx_calendar}
\title{Test whether an object is an \code{idx_calendar}}
\usage{
is_idx_calendar(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical.
}
\description{
Test whether an object is an \code{idx_calendar}
}
