% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_series.R
\name{is_idx_series}
\alias{is_idx_series}
\title{Test whether an object is an \code{idx_series}}
\usage{
is_idx_series(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical.
}
\description{
Test whether an object is an \code{idx_series}
}
