% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_calendar.R
\name{is_multi_step_pattern}
\alias{is_multi_step_pattern}
\title{Test whether an object is a \code{multi_step_pattern}}
\usage{
is_multi_step_pattern(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical.
}
\description{
Test whether an object is a \code{multi_step_pattern}
}
