% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_series.R
\name{length.idx_series}
\alias{length.idx_series}
\title{Number of observations in an \code{idx_series}}
\usage{
\method{length}{idx_series}(x, ...)
}
\arguments{
\item{x}{An \code{idx_series} object.}

\item{...}{Unused.}
}
\description{
Number of observations in an \code{idx_series}
}
