% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapes}
\alias{mapes}
\title{Compute Mean Absolute Percentage Error (MAPE) for Forecasts Against
a Holdout Sample}
\usage{
mapes(res, n.ahead, Y)
}
\arguments{
\item{res}{A \code{FilterResults} or \code{FilterResultsLI} object, obtained from
\code{estimate()} method.}

\item{n.ahead}{Integer specifying the number of periods to forecast
ahead.}

\item{Y}{An \code{idx_series} object containing the original cumulative
dataset.}
}
\value{
A list containing five error metrics for the forecast, with
element names
\itemize{
\item mape: mean absolute percentage error
\item smape: symmetric mean absolute percentage error (between 0 to 100)
\item mae: mean absolute error
\item rmse: root mean squared error
\item coverage: Percentage of holdout sample data points that lie inside
the confidence interval for predictions}
}
\description{
This is a helper function that calculates five error metrics
of a forecast generated by time series growth curve (tsgc) models. It
compares the forecasted values to a holdout sample, providing a measure
of forecast accuracy.
}
