% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{modelKFS}
\alias{modelKFS}
\title{Extract SSModel object within a KFS object}
\usage{
modelKFS(object)
}
\arguments{
\item{object}{A \code{KFS} object.}
}
\value{
The \code{SSModel} object underlying the \code{KFS} fit.
}
\description{
Accessor method to access the fitted \code{SSModel}.
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
model <- SSModelDynamicGompertz$new(Y = gauteng, q = 0.005,
                                    end.date = as.Date("2020-07-20"))
res <- estimate(model)
modelKFS(output(res))

}
