% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_series.R
\name{NCOL.idx_series}
\alias{NCOL.idx_series}
\title{Number of columns in an \code{idx_series} (\code{NCOL} method)}
\usage{
NCOL.idx_series(x)
}
\arguments{
\item{x}{An \code{idx_series} object.}
}
\description{
Provided for convenience when calling \code{NCOL()}
directly on an \code{idx_series}; note that since \code{base::NCOL} is
not an S3 generic, this method is only reached if \code{NCOL} is called
as \code{NCOL.idx_series()} directly, or after \code{NCOL} has been
made generic (e.g. via \code{Matrix} or similar). Prefer
\code{\link{idx_ncol}} for portable code within this package.
}
