% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nintendo_sales}
\alias{nintendo_sales}
\title{Nintendo Sales}
\format{
An object of class \code{"xts"};
\describe{
\item{wii}{Wii sales}
}
}
\usage{
data(nintendo_sales)
}
\description{
Nintendo Sales
}
\examples{
data(nintendo_sales)

}
\references{
Taken from https://vgsales.fandom.com/wiki/Wii
}
\keyword{datasets}
