% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gy_ci}
\alias{plot_gy_ci}
\title{Plots the growth rate of the cumulative variable with confidence
intervals}
\usage{
plot_gy_ci(
  res,
  plt.start = NULL,
  smoothed = FALSE,
  title = NULL,
  series.name = NULL,
  pad.right = NULL,
  axis = NULL
)
}
\arguments{
\item{res}{A \code{FilterResults} or \code{FilterResultsLI} object,
obtained from the \code{estimate()} method.}

\item{plt.start}{Plot start position. \code{NULL} (default) is the
start of the estimation sample.}

\item{smoothed}{Logical value indicating whether to use the smoothed
estimates. Default \code{FALSE} (filtered estimates).}

\item{title}{Title for the plot. \code{NULL} (default) for no title.}

\item{series.name}{Name of the series the growth rate is being computed
for, e.g. \code{'New cases'}.}

\item{pad.right}{Numerical value for the number of positions of blank
space to leave on the right of the graph.}

\item{axis}{An \code{\link{idx_axis_opts}} object controlling x-axis
mode and info box, or \code{NULL} (default) for the previous default
behaviour.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plots the smoothed/filtered growth rate of the difference
in the cumulated variable (\eqn{g_y}) and the associated confidence
intervals.
}
