% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gy_components}
\alias{plot_gy_components}
\title{Plots the growth rate, level and slope of the log cumulative
growth rate}
\usage{
plot_gy_components(
  res,
  plt.start = NULL,
  smoothed = FALSE,
  title = NULL,
  axis = NULL
)
}
\arguments{
\item{res}{A \code{FilterResults} or \code{FilterResultsLI} object,
obtained from the \code{estimate()} method.}

\item{plt.start}{Plot start position. \code{NULL} (default) is the
start of the estimation sample.}

\item{smoothed}{Logical value indicating whether to use the smoothed
estimates of \eqn{\delta} and \eqn{\gamma}. Default \code{FALSE}
(filtered estimates).}

\item{title}{Title for the plot. \code{NULL} (default) for no title.}

\item{axis}{An \code{\link{idx_axis_opts}} object controlling x-axis
mode and info box, or \code{NULL} (default) for the previous default
behaviour.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plots the smoothed/filtered growth rate of the difference
in the cumulated variable (\eqn{g_y}), the smoothed/filtered growth rate
of the cumulated variable (\eqn{g}), and the smoothed/filtered slope of
\eqn{\ln(g)}, \eqn{\gamma}. Following Harvey and Kattuman (2021), we
compute \eqn{g_{y,t}} as \deqn{g_{y,t} = \exp(\delta_t) + \gamma_t.}
}
