% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_log_forecast}
\alias{plot_log_forecast}
\title{Plots forecast and realised values of the log cumulative growth
rate}
\usage{
plot_log_forecast(
  res,
  Y,
  n.ahead = 14,
  plt.start = NULL,
  title = "",
  caption = "",
  axis = NULL
)
}
\arguments{
\item{res}{A \code{FilterResults} or \code{FilterResultsLI} object,
obtained from the \code{estimate()} method.}

\item{Y}{Cumulated dataset (\code{idx_series}) containing future values.}

\item{n.ahead}{Number of positions ahead from the end of the sample to
forecast. Default is 14.}

\item{plt.start}{Plot start position. \code{NULL} (default) is the
start of the estimation sample.}

\item{title}{Plot title.}

\item{caption}{Plot caption.}

\item{axis}{An \code{\link{idx_axis_opts}} object controlling x-axis
mode and info box, or \code{NULL} (default) for the previous default
behaviour. If \code{info_box = TRUE}, the info box is appended below
any \code{caption} already supplied.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plots actual and filtered values of the log cumulative
growth rate (\eqn{\ln(g_t)}) in the estimation sample and the forecast
and realised log cumulative growth rate out of the estimation sample.
}
