% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_rolling_weights}
\alias{plot_rolling_weights}
\title{Rolling weights of Forecast Combination for Dynamic Gompertz and
Leading Indicator Models}
\usage{
plot_rolling_weights(
  Y,
  est.start.date,
  est.end.date,
  all_lags,
  train_days,
  test_days,
  method,
  plot.graph = TRUE,
  LeadIndCol = 1
)
}
\arguments{
\item{Y}{An \code{xts} object with a date index representing the cumulative data series.
The object should contain two columns: the leading indicator and the target variable.
The specific column for the leading indicator can be specified using the \code{LeadIndCol} parameter.}

\item{est.start.date}{A \code{Date} object indicating the start date of the base estimation dataset.
Forecasts generated by the model will include data starting from this date.}

\item{est.end.date}{A \code{Date} object indicating the end date of the base estimation dataset.
The model will use data within the range defined by \code{est.start.date} and \code{est.end.date}
for initial training and estimation. Predictions will then extend beyond \code{est.end.date}
over a period determined by the sum of \code{train_days} and \code{test_days}. Specifically:
\itemize{
\item The period \code{est.end.date + 1 : train_days} is used for training the forecast combination model.
\item The period \code{est.end.date + train_days + 1 : test_days} is used for testing and evaluation.
}}

\item{all_lags}{A numeric vector specifying the lags to be used in the
Leading Indicator model. A 0 value indicates the Dynamic Gompertz Model is used.}

\item{train_days}{An integer specifying the number of days used to train the
forecast combination model.}

\item{test_days}{An integer specifying the number of days used to test the
forecast combination model.}

\item{method}{A function or method from the \code{ForecastComb} package to
compute the optimal
combination of forecasts. Example: \code{comb_OLS}.}

\item{plot.graph}{A logical value. If TRUE, computes the forecast combination
dynamically, showing a graph of the evolution of weights across \code{test_days}.
If FALSE, returns the rolling_comb object. Defaults to TRUE.}

\item{LeadIndCol}{An integer indicating the column number in \code{Y} that
contains the leading indicator. Defaults to 1.}
}
\value{
If \code{plot.graph = FALSE}, returns a ForecastComb object with
details of the computed weights.

If \code{plot.graph = TRUE}, returns a \code{ggplot} object showing the evolution
of weights over time.
}
\description{
This function computes the optimal combination of forecasts from
the Dynamic Gompertz model and the Leading Indicator model over specified lags.
It provides flexibility to train and test the combination model, optionally
displaying the dynamic evolution of weights over time.
}
\examples{
library(tsgc)
library(ForecastComb)
library(tidyr)
Y <- england[, c("cum_cases", "cum_admissions")]
est.start.date <- as.Date("2020-09-01")
est.end.date <- as.Date("2020-10-30")
Y.reinit <- reinitialise_dataframe(Y, est.start.date)

plot_rolling_weights(
  Y.reinit, 
  est.start.date, 
  est.end.date,
  all_lags = c(2, 5, 7, 9), 
  train_days = 20, 
  test_days = 60,
  method = comb_BG
)

}
