% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.FilterResults}
\alias{print.FilterResults}
\title{Calling print method for FilterResults class}
\usage{
\method{print}{FilterResults}(x, ...)
}
\arguments{
\item{x}{A \code{FilterResults} object}

\item{...}{Additional arguments.}
}
\description{
Accessor method to print a short description for the objects of
\code{FilterResults} class
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng, q = 0.005, end.date=as.Date("2020-07-20"))
# Estimate a specified model
res <- estimate(model)

# Return short description of fitted model
print(res)

}
