% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_calendar.R
\name{print.idx_calendar}
\alias{print.idx_calendar}
\title{Print an \code{idx_calendar}}
\usage{
\method{print}{idx_calendar}(x, ...)
}
\arguments{
\item{x}{An \code{idx_calendar} object.}

\item{...}{Unused.}
}
\description{
Print an \code{idx_calendar}
}
