% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_series.R
\name{print.idx_series}
\alias{print.idx_series}
\title{Print an \code{idx_series}}
\usage{
\method{print}{idx_series}(x, ...)
}
\arguments{
\item{x}{An \code{idx_series} object.}

\item{...}{Unused.}
}
\description{
Print an \code{idx_series}
}
