% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_calendar.R
\name{print.idx_step}
\alias{print.idx_step}
\title{Print an \code{idx_step}}
\usage{
\method{print}{idx_step}(x, ...)
}
\arguments{
\item{x}{An \code{idx_step} object.}

\item{...}{Unused.}
}
\description{
Print an \code{idx_step}
}
