% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_calendar.R
\name{print.multi_step_pattern}
\alias{print.multi_step_pattern}
\title{Print a \code{multi_step_pattern}}
\usage{
\method{print}{multi_step_pattern}(x, ...)
}
\arguments{
\item{x}{A \code{multi_step_pattern} object.}

\item{...}{Unused.}
}
\description{
Print a \code{multi_step_pattern}
}
