% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.SSModelDynamicGompertz}
\alias{print.SSModelDynamicGompertz}
\title{Calling print method for SSModelDynamicGompertz class}
\usage{
\method{print}{SSModelDynamicGompertz}(x, ...)
}
\arguments{
\item{x}{A \code{SSModelDynamicGompertz} object}

\item{...}{Additional arguments.}
}
\description{
Accessor method to print a short description for the objects of
\code{SSModelDynamicGompertz} class
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
conv <- xts_to_idx(gauteng)

# Specify a model
model <- SSModelDynamicGompertz$new(Y = conv$series, q = 0.005,
                                    calendar = conv$calendar,
                                    end = idx_to_pos(conv$calendar, "2020-07-06"))

# Print a short description of the model object
print(model)

}
