% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.SSModelLeadingIndicator}
\alias{print.SSModelLeadingIndicator}
\title{Calling print method for classes in tsgc}
\usage{
\method{print}{SSModelLeadingIndicator}(x, ...)
}
\arguments{
\item{x}{A \code{SSModelLeadingIndicator} object}

\item{...}{Additional arguments.}
}
\description{
Accessor method to print a short description for the objects of
\code{SSModelLeadingIndicator} class
}
\examples{
library(tsgc)
data(england, package = "tsgc")
conv <- xts_to_idx(england[, 1:2])

# Specify a model
out_eng <- SSModelLeadingIndicator(
  Y = conv$series, n.lag = 4, sea.period = 7, LeadIndCol = 1,
  calendar = conv$calendar,
  start = idx_to_pos(conv$calendar, "2021-04-30"),
  end = idx_to_pos(conv$calendar, "2021-07-24"))

# Print a short description of the model object
print(out_eng)


}
