% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{Ptt}
\alias{Ptt}
\title{Extract error covariance matrix of filtered states from a KFS object}
\usage{
Ptt(object)
}
\arguments{
\item{object}{A \code{KFS} object.}
}
\value{
The error covariance matrix of the filtered states.
}
\description{
Accessor method to access the non-diffuse part of the error
covariance matrix of the filtered states in a \code{KFS} object.
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
conv <- xts_to_idx(gauteng)
model <- SSModelDynamicGompertz$new(Y = conv$series, q = 0.005,
                                    calendar = conv$calendar,
                                    end = idx_to_pos(conv$calendar, "2020-07-20"))
res <- estimate(model)
Ptt(output(res))
}
