% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qtr2date}
\alias{qtr2date}
\title{Change yearqtr object into date object}
\usage{
qtr2date(dates)
}
\arguments{
\item{dates}{}
}
\description{
This function changes the yearqtr or yearmon object into a date object,
the first day of the quarter in order to allow plotting in ggplot.
}
