% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reinitialise_dataframe}
\alias{reinitialise_dataframe}
\title{Reinitialise a series at a given position}
\usage{
reinitialise_dataframe(dt, reinit.idx)
}
\arguments{
\item{dt}{Cumulated data series, as an \code{idx_series} with exactly one
column.}

\item{reinit.idx}{Integer position at which reinitialisation should
occur (i.e. \eqn{t=r}, using the notation in the vignette).}
}
\value{
The reinitialised series, as an \code{idx_series} starting at
\code{reinit.idx}.
}
\description{
Takes a cumulated series and re-bases it so that it starts
from zero at \code{reinit.idx - 1}.
}
\examples{
x <- idx_series(cumsum(rpois(30, 5)) + 1, start = 1)
reinitialise_dataframe(x, 10)

}
