% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_dates}
\alias{seq_dates}
\title{Regularly spaced time objects}
\usage{
seq_dates(from, resolution, to = NA, length.out = NA)
}
\arguments{
\item{dates}{}
}
\description{
Create a regularly spaced object of class "yearqtr".
}
