% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset.xts}
\alias{subset.xts}
\title{Subsetting xts objects given start dates and end dates}
\usage{
\method{subset}{xts}(df, start.date, end.date = NULL)
}
\arguments{
\item{df}{xts object}

\item{start.date}{Start date of time frame, Date object.}

\item{end.date}{End date of time frame, Date object.}
}
\value{
A subsetted data frame
}
\description{
Helper method to subset a data frame for a specified time frame
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")


}
