% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.FilterResults}
\alias{summary.FilterResults}
\title{Calling summary method for FilterResults}
\usage{
\method{summary}{FilterResults}(object, ...)
}
\arguments{
\item{object}{A \code{FilterResults} object}

\item{...}{Additional arguments.}
}
\description{
Accessor method to show a summary for the objects of
\code{FilterResults} class
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
conv <- xts_to_idx(gauteng)
# Specify a model
model <- SSModelDynamicGompertz$new(Y = conv$series, q = 0.005,
                                    calendar = conv$calendar,
                                    end = idx_to_pos(conv$calendar, "2020-07-20"))
# Estimate a specified model
res <- estimate(model)

# Return KFS object in output of res
summary(res)

}
