% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.FilterResultsLI}
\alias{summary.FilterResultsLI}
\title{Calling summary method for FilterResultsLI}
\usage{
\method{summary}{FilterResultsLI}(object, ...)
}
\arguments{
\item{object}{A \code{FilterResultsLI} object}

\item{...}{Additional arguments.}
}
\description{
Accessor method to show a summary for the objects of
\code{FilterResultsLI} class
}
\examples{
library(tsgc)
data(england, package = "tsgc")
conv <- xts_to_idx(england[, 1:2])

out_eng <- SSModelLeadingIndicator(
  Y = conv$series, n.lag = 4, sea.period = 7, LeadIndCol = 1,
  calendar = conv$calendar,
  start = idx_to_pos(conv$calendar, "2021-04-30"),
  end = idx_to_pos(conv$calendar, "2021-07-24"))

res_eng<-estimate(out_eng)
summary(res_eng)

}
