% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.SSModelDynamicGompertz}
\alias{summary.SSModelDynamicGompertz}
\title{Calling summary method for SSModelDynamicGompertz class}
\usage{
\method{summary}{SSModelDynamicGompertz}(object, ...)
}
\arguments{
\item{object}{A \code{SSModelDynamicGompertz} object}

\item{...}{Additional arguments.}
}
\description{
Accessor method to show a summary for the objects of
\code{SSModelDynamicGompertz} class
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
conv <- xts_to_idx(gauteng)

# Specify a model
model <- SSModelDynamicGompertz$new(Y = conv$series, q = 0.005,
                                    calendar = conv$calendar,
                                    end = idx_to_pos(conv$calendar, "2020-07-06"))

# Show summary of the model object
summary(model)

}
