% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.SSModelLeadingIndicator}
\alias{summary.SSModelLeadingIndicator}
\title{Calling summary method for classes in tsgc}
\usage{
\method{summary}{SSModelLeadingIndicator}(object, ...)
}
\arguments{
\item{object}{A \code{SSModelLeadingIndicator} object}

\item{...}{Additional arguments.}
}
\description{
Accessor method to show a summary for the objects of
\code{SSModelLeadingIndicator} class
}
\examples{
library(tsgc)

# Specify a model
out_eng <- tsgc::SSModelLeadingIndicator(
Y = england[, 1:2], n.lag = 4, sea.period = 7,
start.date = as.Date("2021-04-30"), end.date = as.Date("2021-07-24"))

summary(out_eng)

}
