% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{supply_xpred.new}
\alias{supply_xpred.new}
\title{Supply exogenous predictors for new predictions}
\usage{
supply_xpred.new(object, new.xts, idx = NULL)
}
\arguments{
\item{object}{FilterResults or FilterReusltsLI object}

\item{new.xts}{An xts object containing new exogenous predictors}

\item{idx}{The series number (must be integers 1 or 2) for which exogenous variables are supplied.
Only applicable for FilterResultsLI object. Defaults to NULL.}
}
\description{
Accessor method to provide new exogenous predictors for the
prediction period.
}
\examples{
#FilterResults example
#Load Gauteng weather 
data(gauteng_weather_2021, package = "tsgc")
gauteng_weather<-gauteng_weather_2021[,c(1,3)]

# Set up model and estimate it
model_weather <- SSModelDynamicGompertz$new(Y = cumulative_cases, xpred=gauteng_weather,
                                            start.date=as.Date("2021-02-01"), 
                                           end.date=as.Date("2021-04-19"))
res_weather <- estimate(model_weather)
summary(res_weather)

# Feed future weather data into the results object. Subsetting of gauteng_weather 
#is done inside the function.
supply_xpred.new(res_weather,gauteng_weather)

#FilterResultsLI example
xpred1<-xpred2<-england_weather_2021[,1:4]
mod<-SSModelLeadingIndicator$new(eng, n.lag=4, xpred1=xpred1, xpred2=xpred2, 
                                start.date = as.Date("2021-04-30"), 
                                end.date = as.Date("2021-07-24"))
res_lead.x<-estimate(mod)

supply_xpred.new(res_lead.x,england_weather_2021[,1:4],idx=1)
supply_xpred.new(res_lead.x,england_weather_2021[,1:4],idx=2)

}
