% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_series.R
\name{tail.idx_series}
\alias{tail.idx_series}
\title{Last observations of an \code{idx_series}}
\usage{
\method{tail}{idx_series}(x, n = 6L, ...)
}
\arguments{
\item{x}{An \code{idx_series} object.}

\item{n}{Number of observations to return from the end of \code{x}.
Default \code{6L}. If \code{n} exceeds \code{length(x)}, the full
series is returned.}

\item{...}{Unused.}
}
\value{
An \code{idx_series} of length \code{min(n, length(x))}.
}
\description{
Returns the last \code{n} observations of \code{x}, in
position order. As with \code{\link{head.idx_series}}, \code{n} is a
count of observations, not a set of absolute positions.
}
