% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_daily_ldl}
\alias{add_daily_ldl}
\title{Compute successive increments and log growth rate of 2-variable
cumulated dataset}
\usage{
add_daily_ldl(data, LeadIndCol = 1)
}
\arguments{
\item{data}{Cumulated data series in xts format with date index and 2 columns:
leading indicator and target variable. Can specify which column is leading
indicator by \code{LeadIndCol} parameter.}

\item{LeadIndCol}{Column number of \code{data} that contains the leading
indicator. An integer that can only take values 1 (by default) or 2.}
}
\value{
A data frame with original cumulative variable, successive increments
and log growth rates.
}
\description{
Helper method to compute the successive increments and log
growth rates of cumulated variables. It will compute the successive
increments and log cumulative growth rate for each column in the
2-column data frame, which will then be used to predict or estimate with the
leading indicator model.
}
\examples{
library(tsgc)
data(england,package="tsgc")
add_daily_ldl(england[,c("cum_cases","cum_admissions")],LeadIndCol=1)

}
