% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{alphahat}
\alias{alphahat}
\title{Extract alphahat in SSModel}
\usage{
alphahat(object)
}
\arguments{
\item{object}{A \code{SSModel} object}
}
\description{
Accessor method to access alphahat (smoothed state estimates)
from a fitted \code{SSModel} object which has had the Kalman filter applied to it
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng, q = 0.005, end.date=as.Date("2020-07-20"))
# Estimate a specified model
res <- estimate(model)

#Return alphahat
alphahat(modelKFS(output(res)))

}
