% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{alphahat}
\alias{alphahat}
\title{Extract alphahat in SSModel}
\usage{
alphahat(object)
}
\arguments{
\item{object}{A \code{SSModel} object}
}
\description{
Accessor method to access alphahat in \code{SSModel} object
}
