% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/utils_old.R
\name{argmax}
\alias{argmax}
\title{Return index and value of maximum}
\usage{
argmax(x, decreasing = TRUE)

argmax(x, decreasing = TRUE)
}
\arguments{
\item{x}{Object to have its maximum found, usually an xts object.}

\item{decreasing}{Logical value indicating whether \code{x} should be
ordered in decreasing order. Default is \code{TRUE}. Setting this to
\code{FALSE} would find the minimum.}
}
\value{
The maximum value and its index.

The maximum value and its index.
}
\description{
Similar to Python's argmax function.

Similar to Python's argmax function.
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
argmax(gauteng)
library(tsgc)
data(gauteng,package="tsgc")
argmax(gauteng)
}
