% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{att}
\alias{att}
\title{Extract filtered state estimates used in KFS}
\usage{
att(object)
}
\arguments{
\item{object}{KFS object}
}
\value{
Filtered state estimate in KFS object
}
\description{
Accessor method to access filtered state estimate in KFS object
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
idx.est <- zoo::index(gauteng) <= as.Date("2020-07-20")
# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng[idx.est], q = 0.005)
# Estimate a specified model
res <- estimate(model)

#Return filtered state estimates
att(output(res))

}
