% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cross_val}
\alias{cross_val}
\title{Walk-Forward Validation for Model Comparison Using Mean Absolute
Percentage Error (MAPE)}
\usage{
cross_val(
  Y,
  model_list,
  est.end.date,
  n.ahead = 7,
  n.estimate = 1,
  gap = 1,
  xpred_targ.full = NULL,
  xpred_lead.full = NULL,
  LeadIndCol = 1,
  criterion = "mape"
)
}
\arguments{
\item{Y}{An xts object representing the cumulative data series with a date
index. If a Leading Indicator model is compared, Y should include columns for both the leading indicator and
the target variable. The specific column for the leading indicator can be
designated using the \code{LeadIndCol} parameter.}

\item{model_list}{A list containing \code{SSModelDynamicGompertz} or \code{SSModelLeadingIndicator}
objects, to be compared in a cross validation procedure.}

\item{est.end.date}{The initial estimation end date for model fitting.
Starting from this date, the function re-estimates the model and evaluates
the performance for each lag in \code{all_lags} every \code{gap} days, over a
period of \code{n.estimate} days.}

\item{n.ahead}{Integer specifying the number of days to forecast ahead for
MAPE evaluation.}

\item{n.estimate}{Integer indicating the total number of days for which
walk-forward validation results will be reported.}

\item{gap}{Integer specifying the time gap between two successive validations, where the model
is re-estimated and evaluated during the walk-forward validation.}

\item{xpred_targ.full}{An xts object containing the values of exogenous variables for
the target variable over the estimation and prediction time frame.}

\item{xpred_lead.full}{(Only for required for leading indicator models)
An xts object containing the values of exogenous variables for
the leading indicator over the estimation and prediction time frame.}

\item{LeadIndCol}{(Only required for leading indicator models) Integer
representing the column number in \code{y} that contains the leading
indicator.}

\item{criterion}{A string object indicating how to compare between different
models. Available choices are "mape" (by default), "smape", "mae" and "rmse".}
}
\value{
A table summarizing the chosen error metric for each model in
\code{model_list} across the specified dates.
}
\description{
This function performs a walk-forward validation to compare
forecasting performance across different models specified by the user.
It returns a data frame of a user-specified error metric (e.g. MAPE, MAE)
for forecasts \code{n.ahead} days ahead, using the given models with varying
end dates.
}
\examples{
library(tsgc)
library(KFAS)

#Lay out the estimation settings
est.start <- as.Date("2020-02-25")
est.end <- as.Date("2020-04-01")
Yuk <- tsgc::ukitaly[, "UK"]

# Example 1: ukitaly dataset 
# Create a list to store different models
cv_models<-list()

# Model 1: Vanilla Gompertz
cv_models[["Vanilla_q"]]<-SSModelDynamicGompertz(Y=Yuk, q=0.005, start.date = est.start, end.date = est.end)

# Model 2: Vanilla Gompertz with AR1
cv_models[["Vanilla_ar1"]]<-SSModelDynamicGompertz(Y=Yuk, start.date = est.start, end.date = est.end, ar1=TRUE)

# Model 3-6: Leading Indicator with different lags from 7, 10, 14 or 18
for (i in c(7,10,14,18)){
  cv_models[[paste0("Lag", i)]]<-SSModelLeadingIndicator(Y=ukitaly, start.date = est.start, end.date = est.end, n.lag=i)}

# Display cross-validation analysis
cross_val(Y=ukitaly, model_list=cv_models, est.end.date = est.end, n.estimate=5, gap=2)

# Example 2: England hospitalizations (with xpred)
eng <- tsgc::england[, 1:2]
est.start.eng <- as.Date("2021-04-30")
est.end.eng   <- as.Date("2021-07-24")

#Cross-validation example 
cv_models=list()
# Model 1: Vanilla Gompertz
cv_models[["Vanilla_q"]]<-SSModelDynamicGompertz(Y=eng[,-1], q=0.005, start.date = est.start, end.date = est.end)

# Model 2: Vanilla Gompertz with xpred
cv_models[["Vanilla_xpred"]]<-SSModelDynamicGompertz(Y=eng[,-1], start.date = est.start.eng, end.date = est.end.eng, xpred=england_weather_2021)

# Model 3-6: Leading Indicator with different lags or with xpred
for (i in c(3,4)){
cv_models[[paste0("Lag", i)]]<-SSModelLeadingIndicator(Y=eng, 
start.date = est.start.eng, end.date = est.end.eng, n.lag=i)

cv_models[[paste0("Lag", i,"_xpred")]]<-SSModelLeadingIndicator(Y=eng, 
start.date = est.start.eng, end.date = est.end.eng, 
n.lag=i, xpred_lead = england_weather_2021, xpred_targ=england_weather_2021)}

# Display cross-validation analysis
cross_val(Y=eng, model_list=cv_models, est.end.date = est.end.eng, n.estimate=5, 
gap=2, xpred_targ.full = england_weather_2021, xpred_lead.full = england_weather_2021)

}
