% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cross_val}
\alias{cross_val}
\title{Walk-Forward Validation for Lag Comparison Using Mean Absolute
Percentage Error (MAPE)}
\usage{
cross_val(
  y,
  est.end.date,
  n.ahead,
  all_lags,
  totaldays = 1,
  freq = 1,
  vanilla = TRUE,
  LeadIndCol = 1
)
}
\arguments{
\item{y}{An xts object representing the cumulative data series with a date
index. The object should include columns for both the leading indicator and
the target variable. The specific column for the leading indicator can be
designated using the \code{LeadIndCol} parameter.}

\item{est.end.date}{The initial estimation end date for model fitting.
Starting from this date, the function re-estimates the model and evaluates
the performance for each lag in \code{all_lags} every \code{freq} days, over a
period of \code{totaldays} days.}

\item{n.ahead}{Integer specifying the number of days to forecast ahead for
MAPE evaluation.}

\item{totaldays}{Integer indicating the total number of days for which
walk-forward validation results will be reported.}

\item{freq}{Integer specifying the frequency, in days, at which the model
is re-estimated and evaluated during the walk-forward validation.}

\item{vanilla}{Logical. If \code{TRUE}, the function compares the vanilla
growth curve model to the leading indicator models with different all_lags. The
results for the vanilla model are presented in the row where Lag=0.
Note: If using the vanilla model, the date format must be "\%Y-\%m-\%d".}

\item{LeadIndCol}{(Only required for leading indicator models) Integer
representing the column number in \code{y} that contains the leading
indicator.}
}
\value{
A table summarizing the MAPE scores for each lag across the
specified dates.
}
\description{
This function performs a walk-forward validation to compare
forecasting performance across different lag values. It returns a table of
MAPE values for forecasts \code{n.ahead} days ahead, using models estimated
with varying all_lags over a series of specified end dates.
}
