% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df2ldl}
\alias{df2ldl}
\title{Compute log growth rate of cumulated dataset}
\usage{
df2ldl(dt)
}
\arguments{
\item{dt}{Cumulated data series.}
}
\value{
A data frame of log growth rates of the cumulated variable which has
been inputted via the parameter \code{dt}.
}
\description{
Helper method to compute the log growth rates of cumulated
variables. It will compute the log cumulative growth rate for each column in
the data frame.
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
df2ldl(gauteng)


}
