% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england}
\alias{england}
\title{Cumulative cases of Covid-19 in England.}
\format{
An object of class \code{"xts"};
\describe{
\item{Cases}{Cumulative cases of Covid-19}
}
}
\usage{
data(england)
}
\description{
Cumulative cases of Covid-19 in England.
}
\examples{
data(england)
# plot daily cases
plot(diff(england))
}
\references{
Downloaded from https://ukhsa-dashboard.data.gov.uk/topics/covid-19
}
\keyword{datasets}
