% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england}
\alias{england}
\title{Covid-19 metrics for England.}
\format{
An object of class \code{"xts"};
\describe{
\item{cum_cases}{Cumulative cases of Covid-19}
\item{cum_admissions}{Cumulative hospital admissions with Covid-19 since the start of the pandemic.}
\item{cum_deaths}{Cumulative deaths within 28 days of a positive test for Covid-19 by death date since the start of the pandemic.}
\item{hospital_cases}{Number of patients in hospital with confirmed Covid-19 each day. Note this is NOT cumulative.}
}
}
\usage{
data(england)
}
\description{
Covid-19 metrics for England.
}
\examples{
data(england)
# plot daily cases
plot(diff(england))
}
\references{
Downloaded from https://ukhsa-dashboard.data.gov.uk/topics/covid-19. Now at https://ukhsa-dashboard.data.gov.uk/covid-19-archive-data-download. Full documentation from UKHSA included in downloads from the website.
}
\keyword{datasets}
