% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_weather_2021}
\alias{england_weather_2021}
\title{Daily Weather data of England in 2021.}
\format{
An object of class \code{"xts"} with columns:
\describe{
\item{temperature_C}{Daily mean air temperature in degree Celsius (C) at 2 metres}
\item{max_temp_C}{Daily maximum temperature in degree Celsius (C) at 2 metres}
\item{relhum_percnt}{Relative humidity at 2 metres (\%)}
\item{windspd_mtrs_p_sec}{Wind speed in metres per second (m/s) at 10 metres}
\item{precip_mtrs}{Precipitation (millimetres per day; mm/day)}
}
}
\usage{
data(england_weather_2021)
}
\description{
Daily Weather data of England in 2021.
}
\examples{
data(england_weather_2021)

}
\references{
Downloaded from https://cds.climate.copernicus.eu/datasets/reanalysis-era5-single-levels?tab=overview and aggregated across the latitude and longitude corners for England by the authors.
}
\keyword{datasets}
