% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{estimate}
\alias{estimate}
\title{Calling estimate method for SSModelDynamicGompertz or SSModelLeadingIndicator class}
\usage{
estimate(model)
}
\arguments{
\item{model}{A \code{SSModelDynamicGompertz} or \code{SSModelLeadingIndicator} object}
}
\description{
Accessor method to obtain estimated model for \code{SSModelDynamicGompertz} class
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
idx.est <- zoo::index(gauteng) <= as.Date("2020-07-20")
# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng[idx.est], q = 0.005)
# Estimate a specified model
res <- estimate(model)

}
