% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{estimate}
\alias{estimate}
\title{Calling estimate method for SSModelDynamicGompertz class}
\usage{
estimate(model)
}
\arguments{
\item{model}{A \code{SSModelDynamicGompertz} or \code{SSModelLeadingIndicator} object}
}
\description{
Accessor method to obtain estimated model for \code{SSModelDynamicGompertz} class
}
