% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{estimate}
\alias{estimate}
\title{Calling estimate method for SSModelDynamicGompertz class}
\usage{
estimate(object)
}
\arguments{
\item{object}{SSModelDynamicGompertz object}
}
\description{
Accessor method to obtain estimated model for SSModelDynamicGompertz class
}
