% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_r0}
\alias{estimate_r0}
\title{Calculate reproduction number estimates and credible intervals}
\usage{
estimate_r0(res, gen_int, show_plot = FALSE, title = "Reproduction number")
}
\arguments{
\item{res}{A \code{filterResults} object, obtained from \code{estimate()} method.}

\item{gen_int}{Generation interval in days}

\item{show_plot}{A logical value indicating whether ti show the plot of R0}

\item{title}{Title for the reproduction number plot.}
}
\value{
Forecast of number of periods until peak.
}
\description{
Since Harvey and Kattuman (2021)
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")

}
