% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_r0}
\alias{estimate_r0}
\title{Calculate reproduction number estimates and credible intervals}
\usage{
estimate_r0(
  res,
  gen_int,
  ndays = 7,
  show_plot = FALSE,
  title = "Reproduction number"
)
}
\arguments{
\item{res}{A \code{filterResults} object, obtained from \code{estimate()} method.}

\item{gen_int}{Generation interval in days}

\item{ndays}{Number of days to plot, counting from the end of estimation timeframe.}

\item{show_plot}{A logical value indicating whether ti show the plot of R0}

\item{title}{Title for the reproduction number plot.}
}
\value{
Graph of estimated Rt and forecast intervals.
}
\description{
Following Harvey and Kattuman (2021)
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
cumulative_cases <- gauteng[, 1] 

# Estimate model
model_q <- SSModelDynamicGompertz$new(Y = cumulative_cases, q = 0.005,
                                      start.date=as.Date("2021-02-01"), 
                                      end.date=as.Date("2021-04-19"))
res_q <- estimate(model_q)
summary(res_q)

# Calculate reproduction number estimates and credible intervals.
gen_int <- 4  # Generation interval in days
ndays<-7 #Number of days to plot
r.t <- estimate_r0(res_q, gen_int, ndays)
r.t 

 # Plot reproduction numbers.
estimate_r0(res_q, gen_int, ndays, show_plot = TRUE, title="Gauteng Reproduction numbers")

}
