% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etrading_apps}
\alias{etrading_apps}
\title{Monthly number of downloads of selected leading apps that allow for online
share trading in France from January 2017 to July 2021, by app}
\format{
An object of class \code{"xts"}
}
\usage{
data(etrading_apps)
}
\description{
Monthly number of downloads of selected leading apps that allow for online
share trading in France from January 2017 to July 2021, by app
}
\examples{
data(etrading_apps)

}
\references{
Taken from https://www.statista.com/statistics/1260008/etrading-app-downloads-france/
}
\keyword{datasets}
