% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterResults.R
\docType{class}
\name{FilterResults-class}
\alias{FilterResults-class}
\alias{FilterResults}
\title{FilterResults}
\description{
Class for estimated Dynamic Gompertz Curve model and contains
methods to extract smoothed/filtered estimates of the states, the level of
the incidence variable \eqn{y}, and forecasts of \eqn{y}.
}
\section{Methods}{

\describe{
\item{\code{get_growth_y(smoothed = FALSE, return.components = FALSE)}}{Returns the growth rate of the incidence (\eqn{y}) of the cumulated
variable (\eqn{Y}). Computed as
\deqn{g_t = \exp\{\delta_t\}+\gamma_t.}
 \subsection{Parameters}{\itemize{
  \item{\code{smoothed} Logical value indicating whether to use the
  smoothed estimates of \eqn{\delta} and \eqn{\gamma} to compute the
  growth rate (\code{TRUE}), or the contemporaneous filtered estimates
  (\code{FALSE}). Default is \code{FALSE}.}
  \item{\code{return.components} Logical value indicating whether to
  return the estimates of \eqn{\delta} and \eqn{\gamma} as well as
  the estimates of the growth rate, or just the growth rate. Default is
  \code{FALSE}.}
}}
\subsection{Return Value}{\code{xts} object containing
smoothed/filtered growth rates and components (\eqn{\delta} and
\eqn{\gamma}), where applicable.}}

\item{\code{get_gy_ci(smoothed = FALSE, confidence.level = 0.68)}}{Returns the growth rate of the incidence (\eqn{y}) of the cumulated
variable (\eqn{Y}). Computed as
\deqn{g_t = \exp\{\delta_t\}+\gamma_t.}
 \subsection{Parameters}{\itemize{
  \item{\code{smoothed} Logical value indicating whether to use the
  smoothed estimates of \eqn{\delta} and \eqn{\gamma} to compute the
  growth rate (\code{TRUE}), or the contemporaneous filtered estimates
  (\code{FALSE}). Default is \code{FALSE}.}
  \item{\code{confidence.level} Confidence level for the confidence
  interval.  Default is \eqn{0.68}, which is one standard deviation for
  a normally distributed random variable.}
}}
\subsection{Return Value}{\code{xts} object containing smoothed/filtered
 growth rates and upper and lower bounds for the confidence intervals.}}

\item{\code{initialize(data_xts, index, reinit.date, output)}}{Create an instance of the \code{FilterResults} class.
 \subsection{Parameters}{\itemize{
  \item{\code{data_xts} The cumulated variable.}
  \item{\code{index} The list of dates in the index of \code{data_xts}.}
  \item{\code{reinit.date} The reinitialisation date \eqn{r}. Should
  be specified as an object of class \code{"Date"}.}
  \item{\code{output} A \code{KFS} results object obtained after fitting a 
  \code{SSModelDynamicGompertz} model.}
}}
\subsection{Usage}{\code{SSModelDynGompertzReinit$new(y, q = 0.005,
reinit.date = as.Date("2021-05-12",format = date.format))}}}

\item{\code{mapes(n.ahead, Y)}}{Compute Mean Absolute Percentage Error (MAPE) for trend and seasonal 
forecasts against a holdout sample. For more details, please refer to 
\link{mapes}.}

\item{\code{plot_gy_ci(
  plt.start.date = NULL,
  smoothed = FALSE,
  title = NULL,
  series.name = NULL,
  pad.right = NULL
)}}{Plots the growth rates and the slope of the log cumulative growth rate 
against the dates in estimation sample. 
For more details, please see \link{plot_gy_ci}.}

\item{\code{plot_gy_components(plt.start.date = NULL, smoothed = FALSE, title = NULL)}}{Plots the growth rates and slope of the log cumulative growth rate 
against the dates in estimation sample. 
For more details, please see \link{plot_gy_components}.}

\item{\code{plot_holdout(
  Y,
  n.ahead = 14,
  confidence.level = 0.68,
  date_format = "\%Y-\%m-\%d",
  series.name = "target variable",
  title = NULL,
  caption = NULL
)}}{Plots the forecast of new cases (the difference of the cumulated
variable) over a holdout sample. For more details, please refer to 
\link{plot_holdout}.}

\item{\code{plot_log_forecast(
  Y,
  n.ahead = 14,
  plt.start.date = NULL,
  title = "",
  caption = ""
)}}{Plots actual and filtered values of the log cumulative growth rate 
(\eqn{\ln(g_t)}) in the estimation sample and the forecast and realised 
log cumulative growth rate out of the estimation sample. For more details,
see \link{plot_log_forecast}.}

\item{\code{plot_new_cases(
  n.ahead = 14,
  confidence.level = 0.68,
  date_format = "\%Y-\%m-\%d",
  title = NULL,
  plt.start.date = NULL,
  series.name = "target variable"
)}}{Generates a forecast plot for the difference in the cumulative variable,
showing actual values, forecasts including seasonal components,
and prediction intervals around the forecasts. 
For more details, see \link{plot_new_cases}.}

\item{\code{predict_all(n.ahead, sea.on = TRUE, return.all = FALSE)}}{Returns forecasts of the incidence variable \eqn{y}, the state variables
 and the conditional covariance matrix
for the states.
 \subsection{Parameters}{\itemize{
  \item{\code{n.ahead} The number of forecasts you wish to create from
  the end of your sample period.}
  \item{\code{sea.on} Logical value indicating whether seasonal
  components should be included in the
  state-space model or not. Default is \code{TRUE}.}
  \item{\code{return.all} Logical value indicating whether to return
  all filtered estimates and forecasts
  (\code{TRUE}) or only the forecasts (\code{FALSE}). Default is
  \code{FALSE}.}
}}
\subsection{Return Value}{\code{xts} object containing the forecast
(and filtered, where applicable) level
of \eqn{y} (\code{y.hat}), \eqn{\delta} (\code{level.t.t}),
\eqn{\gamma} (\code{slope.t.t}), vector of states including the
seasonals where applicable (\code{a.t.t}) and covariance matrix of all
states including seasonals where applicable (\code{P.t.t}).}}

\item{\code{predict_level(
  n.ahead,
  confidence.level = 0.68,
  sea.on = FALSE,
  return.diff = TRUE
)}}{Forecast the cumulated variable or the incidence of it. This function returns
the forecast of the cumulated variable \eqn{Y}, or the forecast of the incidence of the cumulated variable, \eqn{y}. For
example, in the case of an epidemic, \eqn{y} might be daily new cases of
the disease and
 \eqn{Y} the cumulative number of recorded infections.
 \subsection{Parameters}{\itemize{
  \item{\code{n.ahead} The number of periods ahead you wish to forecast from
  the end of the estimation window.}
  \item{\code{confidence.level} The confidence level for the log growth
   rate that should be used to compute
  the forecast intervals of \eqn{y}.}
  \item{\code{sea.on} Logical value indicating whether to return the prediction 
  of just the trend or prediction incorporating seasonality.}
  \item{\code{return.diff} Logical value indicating whether to return the cumulated variable,
  \eqn{Y}, or the incidence of it,
  \eqn{y} (i.e., the first difference of the cumulated variable). Default is
  \code{TRUE}.}
}}
\subsection{Return Value}{\code{xts} object containing the point
forecasts and upper and lower bounds of
the forecast interval.}}

\item{\code{print()}}{Provides a quick glimpse of model states and standard errors.}

\item{\code{print_estimation_results()}}{Prints a table of estimated parameters in a format ready to paste into
LaTeX.}

\item{\code{summary()}}{Supplies details of the filterResults object, such as estimated 
parameter values, start and end dates of estimation.}
}}

\examples{
library(tsgc)
data(gauteng,package="tsgc")
idx.est <- zoo::index(gauteng) <= as.Date("2020-07-20")

# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng[idx.est], q = 0.005)
# Estimate a specified model
res <- estimate(model)
# Print estimation results
res$print_estimation_results()
# Forecast 7 days ahead from the end of the estimation window
res$predict_level(y.cum = gauteng[idx.est], n.ahead = 7,
  confidence.level = 0.68)
# Forecast 7 days ahead from the model and return filtered states
res$predict_all(n.ahead = 7, return.all = TRUE)
# Return the filtered growth rate and its components
res$get_growth_y(return.components = TRUE)
# Return smoothed growth rate of incidence variable and its confidence
# interval
res$get_gy_ci(smoothed = TRUE, confidence.level = 0.68)

}
\references{
Harvey, A. C. and Kattuman, P. (2021). A Farewell to R:
Time Series Models for Tracking and
Forecasting Epidemics, Journal of the Royal Society Interface, vol 18(182):
20210179
}
