% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterResultsLI.R
\docType{class}
\name{FilterResultsLI-class}
\alias{FilterResultsLI-class}
\alias{FilterResultsLI}
\title{FilterResultsLI}
\description{
Class for estimated Leading Indicator Gompertz Curve model and
contains methods to extract smoothed/filtered estimates of the states, the
level of the incidence variable \eqn{y}, and forecasts of \eqn{y}.
}
\section{Methods}{

\describe{
\item{\code{get_gy_ci(smoothed = FALSE, confidence.level = 0.68)}}{Returns the growth rate of the incidence (\eqn{y}) of the cumulated
variable (\eqn{Y}). Computed as
\deqn{g_t = \exp\{\delta_t\}+\gamma_t.}
 \subsection{Parameters}{\itemize{
  \item{\code{smoothed} Logical value indicating whether to use the
  smoothed estimates of \eqn{\delta} and \eqn{\gamma} to compute the
  growth rate (\code{TRUE}), or the contemporaneous filtered estimates
  (\code{FALSE}). Default is \code{FALSE}.}
  \item{\code{confidence.level} Confidence level for the confidence
  interval.  Default is \eqn{0.68}, which is one standard deviation for
  a normally distributed random variable.}
}}
\subsection{Return Value}{\code{xts} object containing smoothed/filtered
 growth rates and upper and lower bounds for the confidence intervals.}}

\item{\code{predict_level(n.ahead = n.lag, confidence.level = 0.68)}}{Forecast the cumulated variable or the incidence of it. This function returns
the forecast of the cumulated variable \eqn{Y}, or the forecast of the incidence of the cumulated variable, \eqn{y}. For
example, in the case of an epidemic, \eqn{y} might be daily new cases of
the disease and
 \subsection{Parameters}{\itemize{
  \item{\code{n.ahead} The number of periods ahead you wish to forecast from
  the end of the estimation window. Default is \code{n.lag}.}
  \item{\code{confidence.level} The confidence level for the log growth
   rate that should be used to compute the forecast intervals of \eqn{y}.}
  }
}
\subsection{Return Value}{A list object containing n.lag and 2 \code{xts}
objects: the point forecasts and upper and lower bounds of the forecast interval
for trend and forecast with seasonal component.}}

\item{\code{print_estimation_results()}}{Prints a table of estimated parameters in a format ready to paste into
LaTeX.}
}}

\examples{
library(tsgc)
data(gauteng,package="tsgc")
idx.est <- zoo::index(gauteng) <= as.Date("2020-07-20")

# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng[idx.est], q = 0.005)
# Estimate a specified model
res <- model$estimate()
# Print estimation results
res$print_estimation_results()
# Forecast 7 days ahead from the end of the estimation window
res$predict_level(y.cum = gauteng[idx.est], n.ahead = 7,
  confidence.level = 0.68)
# Forecast 7 days ahead from the model and return filtered states
res$predict_all(n.ahead = 7, return.all = TRUE)
# Return the filtered growth rate and its components
res$get_growth_y(return.components = TRUE)
# Return smoothed growth rate of incidence variable and its confidence
# interval
res$get_gy_ci(smoothed = TRUE, confidence.level = 0.68)

}
\references{
Harvey, A. C. and Kattuman, P. (2021).
}
