% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterResultsLI.R
\docType{class}
\name{FilterResultsLI-class}
\alias{FilterResultsLI-class}
\alias{FilterResultsLI}
\title{Class for the estimated Leading Indicator Gompertz Curve model}
\description{
Class for estimated Leading Indicator Gompertz Curve model and
contains methods to extract smoothed/filtered estimates of the states, the
level of the incidence variable \eqn{y}, and forecasts of \eqn{y}.
}
\section{Fields}{

\describe{
\item{\code{data_xts}}{A xts object with cumulated variables: lagged leading indicator and
target variable.}

\item{\code{output}}{A \code{KFS} results object obtained after fitting a
\code{SSModelLeadingIndicator} model.}

\item{\code{n.lag}}{Number of days to lag the leading indicator, inherited from the
estimated \code{SSModelLeadingIndicator} model.}

\item{\code{sea.period}}{The period of seasonality, inherited from the estimated
\code{SSModelLeadingIndicator} model. For a day-of-the-week
effect with daily data, this would be 7.}

\item{\code{LeadIndCol}}{The column in \code{data_xts} that contains the leading
indicator, inherited from the estimated \code{SSModelLeadingIndicator} model.}

\item{\code{xpred_logical}}{Vector of length 2 with logical values, indicating whether
there are exogenous predictors for leading series and target series.}

\item{\code{xpred1.new}}{An xts object containing the values of exogenous variables for
the leading indicator over the prediction time frame.}

\item{\code{xpred2.new}}{An xts object containing the values of exogenous variables for
the target variable over the prediction time frame.}

\item{\code{resolution}}{A character object showing the time resolution of the data
in \code{data_xts}. Options are "daily", "monthly, "quarterly" and "yearly".
Automatically estimated when \code{data_xts} is provided.}
}}

\section{Methods}{

\describe{
\item{\code{get_growth_y(smoothed = FALSE, return.components = FALSE)}}{Returns the growth rate of the incidence (\eqn{y}) of the cumulated
variable (\eqn{Y}). Computed as
\deqn{g_t = \exp\{\delta_t\}+\gamma_t.}
 \subsection{Parameters}{\itemize{
  \item{\code{smoothed} Logical value indicating whether to use the
  smoothed estimates of \eqn{\delta} and \eqn{\gamma} to compute the
  growth rate (\code{TRUE}), or the contemporaneous filtered estimates
  (\code{FALSE}). Default is \code{FALSE}.}
  \item{\code{return.components} Logical value indicating whether to
  return the estimates of \eqn{\delta} and \eqn{\gamma} as well as
  the estimates of the growth rate, or just the growth rate. Default is
  \code{FALSE}.}
}}
\subsection{Return Value}{\code{xts} object containing
smoothed/filtered growth rates and components (\eqn{\delta} and
\eqn{\gamma}), where applicable.}}

\item{\code{get_gy_ci(smoothed = FALSE, confidence.level = 0.68)}}{Returns the growth rate of the incidence (\eqn{y}) of the cumulated
variable (\eqn{Y}). Computed as
\deqn{g_t = \exp\{\delta_t\}+\gamma_t.}
 \subsection{Parameters}{\itemize{
  \item{\code{smoothed} Logical value indicating whether to use the
  smoothed estimates of \eqn{\delta} and \eqn{\gamma} to compute the
  growth rate (\code{TRUE}), or the contemporaneous filtered estimates
  (\code{FALSE}). Default is \code{FALSE}.}
  \item{\code{confidence.level} Confidence level for the confidence
  interval.  Default is \eqn{0.68}, which is one standard deviation for
  a normally distributed random variable.}
}}
\subsection{Return Value}{\code{xts} object containing smoothed/filtered
 growth rates and upper and lower bounds for the confidence intervals.}}

\item{\code{initialize(
  data_xts,
  output,
  n.lag,
  sea.period,
  LeadIndCol,
  xpred_logical,
  start.date,
  end.date,
  xpred1.new = NULL,
  xpred2.new = NULL,
  resolution = "daily"
)}}{Create an instance of the \code{FilterResultsLI} class with fields defined
earlier in the fields section.}

\item{\code{mapes(n.ahead, Y)}}{Compute Mean Absolute Percentage Error (MAPE) for trend and seasonal 
forecasts against a holdout sample. For more details, please refer to 
\link{mapes}.}

\item{\code{plot_forecast(
  n.ahead = 7,
  confidence.level = 0.68,
  title = NULL,
  plt.start.date = NULL,
  series.name = "target variable"
)}}{Generates a forecast plot for the difference in the cumulative target 
variable, showing actual values, forecasts including seasonal components,
and prediction intervals around the forecasts. For more details, see 
\link{plot_forecast}.}

\item{\code{plot_gy_ci(
  plt.start.date = NULL,
  smoothed = FALSE,
  title = NULL,
  series.name = NULL,
  pad.right = NULL
)}}{Plots the growth rates and the slope of the log cumulative growth rate of 
the target variable against the dates in estimation sample. 
  For more details, please see \link{plot_gy_ci}.}

\item{\code{plot_gy_components(plt.start.date = NULL, smoothed = FALSE, title = NULL)}}{Plots the growth rates and slope of the log cumulative growth rate 
against the dates in estimation sample. 
For more details, please see \link{plot_gy_components}.}

\item{\code{plot_holdout(
  Y,
  n.ahead = 14,
  confidence.level = 0.68,
  series.name = "target variable",
  title = NULL,
  caption = NULL
)}}{Plots the forecast of the target variable over a holdout sample. 
For more details, please refer to \link{plot_holdout}.}

\item{\code{plot_log_forecast(
  Y,
  n.ahead = 14,
  plt.start.date = NULL,
  title = "",
  caption = ""
)}}{Plots actual and filtered values of the log cumulative growth rate 
(\eqn{\ln(g_t)}) of the target variable in the estimation sample and 
the forecast and realised log cumulative growth rate of the target variable
out of the estimation sample. For more details, see \link{plot_log_forecast}.}

\item{\code{predict_all(
  n.ahead,
  sea.on = FALSE,
  return.all = FALSE,
  confidence.level = 0.68
)}}{Returns forecasts of the incidence variable \eqn{y}, the state variables
 and the conditional covariance matrix
for the states.
 \subsection{Parameters}{\itemize{
  \item{\code{n.ahead} The number of forecasts you wish to create from
  the end of your sample period.}
  \item{\code{sea.on} Logical value indicating whether seasonal
  components should be included in the
  state-space model or not. Default is \code{FALSE}.}
  \item{\code{return.all} Logical value indicating whether to return
  all filtered estimates and forecasts
  (\code{TRUE}) or only the forecasts (\code{FALSE}). Default is
  \code{FALSE}.}
}}
\subsection{Return Value}{\code{xts} object containing the forecast
(and filtered, where applicable) level
of \eqn{y} (\code{y.hat}), \eqn{\delta} (\code{level.t.t}),
\eqn{\gamma} (\code{slope.t.t}), vector of states including the
seasonals where applicable (\code{a.t.t}) and covariance matrix of all
states including seasonals where applicable (\code{P.t.t}).}}

\item{\code{predict_level(n.ahead = n.lag, confidence.level = 0.68, sea.on = FALSE)}}{Forecast the cumulated variable or the incidence of it. This function returns
the forecast of the cumulated variable \eqn{Y}, or the forecast of the incidence of the cumulated variable, \eqn{y}. For
example, in the case of an epidemic, \eqn{y} might be daily new cases of
the disease and
 \subsection{Parameters}{\itemize{
  \item{\code{n.ahead} The number of periods ahead you wish to forecast from
  the end of the estimation window. Default is \code{n.lag}.}
  \item{\code{sea.on} Logical value indicating whether to return the prediction 
  of just the trend or prediction incorporating seasonality.}
  \item{\code{confidence.level} The confidence level for the log growth
   rate that should be used to compute the forecast intervals of \eqn{y}.}
 }
}
\subsection{Return Value}{A list object containing n.lag and 2 \code{xts}
objects: the point forecasts and upper and lower bounds of the forecast interval
for trend and forecast with seasonal component.}}

\item{\code{print()}}{Provides a quick glimpse of model states and standard errors.}

\item{\code{print_estimation_results()}}{Prints a table of estimated parameters in a format ready to paste into
LaTeX.}

\item{\code{summary()}}{Supplies details of the filterResults object, such as estimated 
parameter values, start and end dates of estimation.}
}}

\examples{
library(tsgc)

#Define estimation environment
Y = england[,1:2] #select only cumulative cases and cumulative hospitalizations
estimation.date.start = as.Date("2021-04-30")
estimation.date.end = as.Date("2021-07-24")
plt.length=30
q = NULL
confidence.level = 0.68
plt.length = 14
n.lag = 4

# Define and Estimate the model
out<-SSModelLeadingIndicator(Y=Y,n.lag = n.lag,q=q,LeadIndCol=1, sea.period=7,
start.date=estimation.date.start, end.date=estimation.date.end) 
res<-estimate(out) 

# Print estimation results
res$print_estimation_results()

# Forecast 7 days ahead from the end of the estimation window
res$predict_level(n.ahead = 7, confidence.level = 0.68)
  
# Forecast 7 days ahead from the model and return filtered states
res$predict_all(n.ahead = 7, return.all = TRUE)

# Return the filtered growth rate and its components
res$get_growth_y(return.components = TRUE)

# Return smoothed growth rate of incidence variable and its confidence
# interval
res$get_gy_ci(smoothed = TRUE, confidence.level = 0.68)

# Plot forecast of new cases 7 days ahead
res$plot_log_forecast(Y=Y, n.ahead=7,
plt.start.date=estimation.date.end-plt.length)

# Plot forecast of new cases 7 days ahead
res$plot_forecast(n.ahead=7,
plt.start.date = estimation.date.end-plt.length,
series.name="hospitalizations")

# Plot forecasts and outcomes over evaluation period
res$plot_holdout(Y=Y,n.ahead=7, series.name="hospitalizations")

# Plot filtered gy, g and gamma
res$plot_gy_components(plt.start.date = estimation.date.end-plt.length)

# Plot filtered gy, g and gamma
res$plot_gy_ci(plt.start.date = estimation.date.end-plt.length)

#Return MAPE of forecast
res$mapes(n.ahead=7,Y)

}
\references{
Harvey, A. (2021). TIME SERIES MODELLING OF EPIDEMICS:
LEADING INDICATORS, CONTROL GROUPS AND POLICY ASSESSMENT.
National Institute Economic Review, 257, 83–100.
doi:10.1017/nie.2021.21
}
