% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gauteng_weather}
\alias{gauteng_weather}
\title{Weather data of the South African province of Gauteng.}
\format{
An object of class \code{"xts"};
\describe{
\item{TempC}{Temperature in degree Celsius}
\item{TempC_max}{Temperature in degree Celsius in a day}
\item{RelHumid}{Relative humidity}
\item{WindSpeed}{Wind speed}
}
}
\usage{
data(gauteng_weather)
}
\description{
Weather data of the South African province of Gauteng.
}
\examples{
data(gauteng_weather)
# plot daily cases
plot(diff(gauteng_weather))
}
\references{
Downloaded from NASA POWER
}
\keyword{datasets}
