% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gauteng_weather_2021}
\alias{gauteng_weather_2021}
\title{Daily weather data of the South African province of Gauteng in 2021}
\format{
An object of class \code{"xts"} with columns:
\describe{
\item{TempC}{Mean air temperature in degree Celsius at 2 meters}
\item{TempC_max}{Maximum Temperature in degree Celsius at 2 meters}
\item{RelHumid}{Relative humidity (\%)}
\item{WindSpeed}{Wind speed in m/s at 10 meters }
}
}
\usage{
data(gauteng_weather_2021)
}
\description{
Daily weather data of the South African province of Gauteng in 2021
}
\examples{
data(gauteng_weather_2021)
}
\references{
Downloaded from https://cds.climate.copernicus.eu/datasets/reanalysis-era5-single-levels?tab=overview and aggregated across the latitude and longitude corners for Gauteng province, South Africa by the authors.
}
\keyword{datasets}
