% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_time_resolution}
\alias{get_time_resolution}
\title{Identify time resolution of given dates}
\usage{
get_time_resolution(dates)
}
\arguments{
\item{dates}{A vector of dates}
}
\value{
A character string from "daily", "monthly, "quarterly" and "yearly".
}
\description{
This function identifies the time resolution of a vector of
dates, where all dates must come from the same class ("yearmon", "yearqtr",
"date").
}
