% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/utils_old.R
\name{get_timeframe}
\alias{get_timeframe}
\title{Subsetting xts objects given start dates and end dates}
\usage{
get_timeframe(df, start.date, end.date = NULL)

get_timeframe(df, start.date, end.date = NULL)
}
\arguments{
\item{df}{An xts object, or NULL if no data are supplied.}

\item{start.date}{Start date of time frame, Date object.}

\item{end.date}{End date of time frame, Date object.}
}
\value{
An xts object containing the selected observations, or
\code{NULL} if \code{df} is \code{NULL}.

An xts object containing the selected observations, or
\code{NULL} if \code{df} is \code{NULL}.
}
\description{
Helper method to subset a data frame for a specified time frame

Helper method to subset a data frame for a specified time frame
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")

# Select data between 2020-04-01 and 2020-08-01, inclusive
get_timeframe(gauteng, as.Date("2020-04-01"), as.Date("2020-08-01"))

# Select all data after 2020-04-01, inclusive
get_timeframe(gauteng, as.Date("2020-04-01"))

library(tsgc)
data(gauteng,package="tsgc")

# Select data between 2020-04-01 and 2020-08-01, inclusive
get_timeframe(gauteng, as.Date("2020-04-01"), as.Date("2020-08-01"))

# Select all data after 2020-04-01, inclusive
get_timeframe(gauteng, as.Date("2020-04-01"))

}
