% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_timeframe}
\alias{get_timeframe}
\title{Subsetting xts objects given start dates and end dates}
\usage{
get_timeframe(df, start.date, end.date = NULL)
}
\arguments{
\item{df}{xts object}

\item{start.date}{Start date of time frame, Date object.}

\item{end.date}{End date of time frame, Date object.}
}
\value{
A subsetted data frame
}
\description{
Helper method to subset a data frame for a specified time frame
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")

# Select data between 2020-04-01 and 2020-08-01, inclusive
get_timeframe(gauteng, as.Date("2020-04-01"), as.Date("2020-08-01"))

# Select all data after 2020-04-01, inclusive
get_timeframe(gauteng, as.Date("2020-04-01"))

}
