% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{get_V}
\alias{get_V}
\title{Extract error covariance matrix of smoothed states from a KFS object}
\usage{
get_V(object)
}
\arguments{
\item{object}{A \code{KFS} object.}
}
\value{
The error covariance matrix of the smoothed states.
}
\description{
Accessor method to access the non-diffuse part of the error
covariance matrix of the smoothed states in a \code{KFS} object.
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
model <- SSModelDynamicGompertz$new(Y = gauteng, q = 0.005,
                                    end.date = as.Date("2020-07-20"))
res <- estimate(model)
get_V(output(res))

}
