% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{gety.hat}
\alias{gety.hat}
\title{Extract prediction y.hat in SSModel}
\usage{
gety.hat(object)
}
\arguments{
\item{object}{An \code{SSModel} object}
}
\description{
Accessor method to access prediction y.hat in \code{SSModel} object
}
