% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{gety.hat}
\alias{gety.hat}
\title{Extract prediction y.hat in predict.all output}
\usage{
gety.hat(object)
}
\arguments{
\item{object}{An object from the output of predict.all}
}
\description{
Accessor method to access prediction y.hat in \code{SSModel} object
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")

# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng, q = 0.005, 
                                    end.date=as.Date("2020-07-20"))

# Estimate a specified model
res <- estimate(model)

# Get object from predict.all
all_predictions<-res$predict_all(n.ahead=7)

# Get prediction
gety.hat(all_predictions)

}
