% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{gety}
\alias{gety}
\title{Extract time series y from an SSModel object}
\usage{
gety(object)
}
\arguments{
\item{object}{An \code{SSModel} object.}
}
\value{
The time series \code{y}.
}
\description{
Accessor method to access the time series \code{y} in an
\code{SSModel} object.
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
conv <- xts_to_idx(gauteng)
model <- SSModelDynamicGompertz$new(Y = conv$series, q = 0.005,
                                    calendar = conv$calendar,
                                    end = idx_to_pos(conv$calendar, "2020-07-20"))
res <- estimate(model)
gety(modelKFS(output(res)))
}
