% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{gety}
\alias{gety}
\title{Extract time series y in SSModel}
\usage{
gety(object)
}
\arguments{
\item{object}{SSModel object}
}
\description{
Accessor method to access time series y in \code{SSModel} object
}
