% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapes}
\alias{mapes}
\title{Compute Mean Absolute Percentage Error (MAPE) for Forecasts Against
a Holdout Sample}
\usage{
mapes(object, ...)
}
\arguments{
\item{object}{A \code{filterResults} or \code{filterResultsLI} object, obtained from
\code{estimate()} method.}

\item{n.ahead}{Integer specifying the number of days to forecast ahead.}

\item{Y}{An xts object containing the original cumulative dataset
with two columns: the leading indicator and the target variable.}
}
\value{
A list containing the MAPE values for both the trend forecast and
the forecast that includes the seasonal component.
}
\description{
This is a helper function that calculates the Mean Absolute
Percentage Error (MAPE) of a forecast generated by time series growth curve
(tsgc) models. It compares the forecasted values to a holdout sample,
providing a measure of forecast accuracy. If the vanilla growth curve model
is used, ensure that the date format is "\%Y-\%m-\%d".
}
\examples{
date_format="\%Y-\%m-\%d"
est.end.date=as.Date("2021-07-24")

model_q <- SSModelDynamicGompertz$new(Y = gauteng[index(gauteng) <= est.end.date])
res <- estimate(model_q)

mapes(res,n.ahead,model_q)


}
