% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapes}
\alias{mapes}
\title{Compute Mean Absolute Percentage Error (MAPE) for Forecasts Against
a Holdout Sample}
\usage{
mapes(res, n.ahead, Y)
}
\arguments{
\item{res}{A \code{filterResults} or \code{filterResultsLI} object, obtained from
\code{estimate()} method.}

\item{n.ahead}{Integer specifying the number of days to forecast ahead.}

\item{Y}{An xts object containing the original cumulative dataset.}
}
\value{
A list containing the MAPE values for both the trend forecast and
the forecast that includes the seasonal component.
}
\description{
This is a helper function that calculates the Mean Absolute
Percentage Error (MAPE) of a forecast generated by time series growth curve
(tsgc) models. It compares the forecasted values to a holdout sample,
providing a measure of forecast accuracy.
}
\examples{
library(tsgc)

#Setup
date_format="\%Y-\%m-\%d"
estimation.date.start=as.Date("2021-04-30")
estimation.date.end=as.Date("2021-07-24")
n.ahead=7
Y=gauteng
idx.est =(zoo::index(Y) >= estimation.date.start) & (zoo::index(Y) <= estimation.date.end)
y = Y[idx.est]

#Estimate the model
model_q <- SSModelDynamicGompertz$new(Y = y)
res <- estimate(model_q)

#Return MAPE of forecast
mapes(res,n.ahead=n.ahead,Y)


}
