% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{modelKFS}
\alias{modelKFS}
\title{Extract SSModel object within KFS object}
\usage{
modelKFS(object)
}
\arguments{
\item{object}{KFS object}
}
\description{
Accessor method to access the fitted SSModel
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")

# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng, end.date=as.Date("2020-07-06"), q = 0.005)
# Estimate a specified model
res <- model$estimate()

# Extract Z matrix from output(res)
modelKFS(output(res))

}
