% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{modelKFS}
\alias{modelKFS}
\title{Extract SSModel object within KFS object}
\usage{
modelKFS(object)
}
\arguments{
\item{object}{KFS object}
}
\description{
Accessor method to access the fitted SSModel
}
