% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nintendo_sales}
\alias{nintendo_sales}
\title{Quarterly Sales of Nintendo consoles from 2004 Q4 to 2024 Q1}
\format{
An object of class \code{"xts"} with columns:
\describe{
\item{wii}{Cumulative Wii sales}
\item{wiiu}{Cumulative Wii U sales}
\item{switch_all}{Cumulative Switch sales (original, lite and OLED combined)}
\item{switch_base}{Cumulative original Switch sales}
\item{switch_Lite}{Cumulative Switch Lite sales}
\item{switch_OLED}{Cumulative switch OLED sales}
\item{3ds}{Cumulative 3DS sales}
\item{nds}{Cumulative NDS sales}
}
}
\usage{
data(nintendo_sales)
}
\description{
Quarterly Sales of Nintendo consoles from 2004 Q4 to 2024 Q1
}
\examples{
data(nintendo_sales)

}
\references{
Taken from https://vgsales.fandom.com/wiki/Wii
}
\keyword{datasets}
