% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{output}
\alias{output}
\title{Extract output of FilterResults or FilterResultsLI}
\usage{
output(object)
}
\arguments{
\item{object}{A \code{FilterResults} or \code{FilterResultsLI} object.}
}
\value{
The fitted KFS model.
}
\description{
Accessor method to access the fitted KFS model from
\code{FilterResults} or \code{FilterResultsLI}.
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
model <- SSModelDynamicGompertz$new(Y = gauteng, q = 0.005,
                                    end.date = as.Date("2020-07-20"))
res <- estimate(model)
output(res)

}
