% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{output}
\alias{output}
\title{Extract output of FilterResults}
\usage{
output(object)
}
\arguments{
\item{object}{FilterResults object}
}
\value{
The fitted KFS model
}
\description{
Accessor method to access the fitted KFS model from \code{FilterResults}
}
