% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{plot.SSModelDynamicGompertz}
\alias{plot.SSModelDynamicGompertz}
\title{Calling plot method for SSModelDynamicGompertz class}
\usage{
\method{plot}{SSModelDynamicGompertz}(
  x,
  y,
  ...,
  title = NULL,
  series.name = "target variable",
  date_break = NULL,
  MA_period = 7
)
}
\arguments{
\item{x}{A \code{SSModelDynamicGompertz} object.}

\item{y}{Unused. Required to match generic signature.}

\item{...}{Additional arguments.}

\item{title}{Title for forecast plot.}

\item{series.name}{The name of the series the growth rate is being computed for.}

\item{date_break}{A character string specifying the interval between date labels.}

\item{MA_period}{Number of days used in centered moving average.}
}
\description{
Accessor method to call the plot method of an object of
\code{SSModelDynamicGompertz} class
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
idx.est <- zoo::index(gauteng) <= as.Date("2020-07-06")

# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng[idx.est], q = 0.005)

# Show summary of the model object
plot(model)

}
