% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{plot.SSModelDynamicGompertz}
\alias{plot.SSModelDynamicGompertz}
\title{Calling plot method for SSModelDynamicGompertz class}
\usage{
\method{plot}{SSModelDynamicGompertz}(
  model,
  title = NULL,
  series.name = "target variable",
  date_break = NULL,
  MA_period = 7
)
}
\arguments{
\item{model}{A \code{SSModelDynamicGompertz} object}
}
\description{
Accessor method to call the plot method of an object of
\code{SSModelDynamicGompertz} class
}
