% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{plot.SSModelLeadingIndicator}
\alias{plot.SSModelLeadingIndicator}
\title{Calling plot method for \code{SSModelLeadingIndicator} class}
\usage{
\method{plot}{SSModelLeadingIndicator}(
  x,
  y,
  ...,
  title = NULL,
  series.name.lead = "Leading Indicator",
  series.name.target = "Target Variable",
  date_break = NULL,
  take.log = TRUE
)
}
\arguments{
\item{x}{A \code{SSModelLeadingIndicator} object.}

\item{y}{Unused. Required to match generic signature.}

\item{...}{Additional arguments.}

\item{title}{Title for forecast plot.}

\item{series.name.lead}{The name of the leading indicator series.}

\item{series.name.target}{The name of the target variable series.}

\item{date_break}{A character string specifying the interval between date labels.}

\item{take.log}{A logical value indicating whether to take the log of the lagged differences.}
}
\description{
Accessor method to call the plot method of an object of
\code{SSModelLeadingIndicator} class
}
\examples{
library(tsgc)

# Specify a model
out_eng <- tsgc::SSModelLeadingIndicator(
Y = england[, 1:2], n.lag = 4, sea.period = 7,
start.date = as.Date("2021-04-30"), end.date = as.Date("2021-07-24"))

plot(out_eng)

}
