% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{plot.SSModelLeadingIndicator}
\alias{plot.SSModelLeadingIndicator}
\title{Calling plot method for classes in tsgc}
\usage{
\method{plot}{SSModelLeadingIndicator}(
  model,
  title = NULL,
  series.name.lead = "Leading Indicator",
  series.name.target = "Target Variable",
  date_break = "60 days",
  take.log = TRUE
)
}
\arguments{
\item{model}{A \code{SSModelLeadingIndicator} object}
}
\description{
Accessor method to call the plot method of an object of
\code{SSModelLeadingIndicator} class
}
