% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_compare_forecast}
\alias{plot_compare_forecast}
\title{Forecast comparison plot}
\usage{
plot_compare_forecast(
  results,
  n.ahead = 14,
  sea.on = TRUE,
  actual = NULL,
  title = "Comparison of forecasts"
)
}
\arguments{
\item{results}{A list of \code{filterResults} or \code{filterResultsLI} object, obtained from
\code{estimate()} method.}

\item{n.ahead}{The duration of the holdout sample. Default is 14.}

\item{sea.on}{Logical value indicating whether to plot the seasonality-adjusted
forecasts. Defaults to \code{TRUE}.}

\item{actual}{Actual values of the cumulated variable, can be the raw dataset.
Subsetting is done within the code.}

\item{title}{Title for forecast plot. Enter as text string. \code{NULL}
(i.e. no title) by default.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plots the forecasts of the cumulated variable (both including and excluding the
seasonal component, where a seasonal is specified) for selected models
and optionally the actual outcomes from the holdout sample. The
forecast intervals are based on the prediction intervals for \eqn{\ln(g_t)}.
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
idx.est <- zoo::index(gauteng) <= as.Date("2020-07-20")
n.ahead=7


}
