% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_compare_forecast}
\alias{plot_compare_forecast}
\title{Forecast comparison plot}
\usage{
plot_compare_forecast(
  results,
  n.ahead = 14,
  sea.on = TRUE,
  actual = NULL,
  title = "Comparison of forecasts"
)
}
\arguments{
\item{results}{A list of \code{FilterResults} or \code{FilterResultsLI} object, obtained from
\code{estimate()} method.}

\item{n.ahead}{The duration of the holdout sample. Default is 14.}

\item{sea.on}{Logical value indicating whether to plot the seasonality-adjusted
forecasts. Defaults to \code{TRUE}.}

\item{actual}{Time series of actual cumulative values. The function
automatically extracts the observations corresponding to the prediction
period, so the complete series may be provided.}

\item{title}{Title for forecast plot. Enter as text string. Defaults as "Comparison of forecasts".}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plots forecasts from a list of fitted models on the same axes for visual comparison.
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
end.date<- as.Date("2020-07-20")

# Model 1: The signal-to-noise ratio is a free parameter
model <- SSModelDynamicGompertz$new(Y = gauteng, end.date=end.date)
res <- estimate(model)

# Model 2: The signal-to-noise ratio is fixed
model_q <- SSModelDynamicGompertz$new(Y = gauteng, end.date=end.date, q=0.005)
res_q <- estimate(model_q)

# Compare forecast
plot_compare_forecast(list(res,res_q), n.ahead=7)

}
