% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plots the forecast of new cases (the difference of the cumulated
variable)}
\usage{
plot_forecast(
  res,
  n.ahead = 7,
  confidence.level = 0.68,
  title = NULL,
  plt.start.date = NULL,
  series.name = "target variable"
)
}
\arguments{
\item{res}{A \code{FilterResults} or \code{FilterResultsLI} object, obtained from
\code{estimate()} method.}

\item{n.ahead}{Number of forecasts (i.e. number of periods ahead to forecast
from end of estimation window). Default is 14.}

\item{confidence.level}{Width of prediction interval for \eqn{\ln g_t} to
use in forecasts of \eqn{y_t = \Delta Y_t}. Default is 0.68, which is
approximately one standard deviation for a Normal distribution.}

\item{title}{Title for forecast plot. Enter as text string. \code{NULL}
(i.e. no title) by default.}

\item{plt.start.date}{First date of actual data (from estimation sample) to
plot on graph.\code{NULL} (i.e. plots all data in estimation window) by
default.}

\item{series.name}{The name of the series the growth rate is being computed
for. E.g. \code{'cases'}. Default is "target variable".}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plots actual values of the difference in the cumulated variable,
the forecasts of the cumulated variable (both including and excluding the
seasonal component, where a seasonal is specified) and forecast intervals
around the forecasts. The forecast intervals are based on the prediction
intervals for \eqn{\ln(g_t)}.
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")

# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng, q = 0.005, end.date=as.Date("2020-07-20"))
# Estimate a specified model
res <- estimate(model)

# Plot forecast of new cases 7 days ahead
plot_forecast(res, n.ahead=7, confidence.level = 0.68, 
title = "Forecast new cases", plt.start.date = as.Date("2020-07-13"),
series.name="cases")

}
