% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.FilterResults}
\alias{print.FilterResults}
\title{Calling print method for FilterResults class}
\usage{
\method{print}{FilterResults}(x, ...)
}
\arguments{
\item{x}{A \code{FilterResults} object}

\item{...}{Additional arguments.}
}
\description{
Accessor method to print a short description for the objects of
\code{FilterResults} class
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
conv <- xts_to_idx(gauteng)
# Specify a model
model <- SSModelDynamicGompertz$new(Y = conv$series, q = 0.005,
                                    calendar = conv$calendar,
                                    end = idx_to_pos(conv$calendar, "2020-07-20"))
# Estimate a specified model
res <- estimate(model)

# Return short description of fitted model
print(res)

}
