% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{print.FilterResults}
\alias{print.FilterResults}
\title{Calling print method for FilterResults class}
\usage{
\method{print}{FilterResults}(model)
}
\arguments{
\item{model}{A \code{FilterResults} object}
}
\description{
Accessor method to print a short description for the objects of
\code{FilterResults} class
}
