% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.FilterResultsLI}
\alias{print.FilterResultsLI}
\title{Calling print method for FilterResultsLI class}
\usage{
\method{print}{FilterResultsLI}(x, ...)
}
\arguments{
\item{x}{A \code{FilterResultsLI} object}

\item{...}{Additional arguments.}
}
\description{
Accessor method to print a short description for the objects of
\code{FilterResultsLI} class
}
\examples{
library(tsgc)
data(england, package = "tsgc")
conv <- xts_to_idx(england[, 1:2])

out_eng <- SSModelLeadingIndicator(
  Y = conv$series, n.lag = 4, sea.period = 7, LeadIndCol = 1,
  calendar = conv$calendar,
  start = idx_to_pos(conv$calendar, "2021-04-30"),
  end = idx_to_pos(conv$calendar, "2021-07-24"))

res_eng<-estimate(out_eng)
print(res_eng) 

}
