% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{print.FilterResultsLI}
\alias{print.FilterResultsLI}
\title{Calling print method for FilterResultsLI class}
\usage{
\method{print}{FilterResultsLI}(model)
}
\arguments{
\item{model}{A \code{FilterResultsLI} object}
}
\description{
Accessor method to print a short description for the objects of
\code{FilterResultsLI} class
}
