% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{print.SSModelDynamicGompertz}
\alias{print.SSModelDynamicGompertz}
\title{Calling print method for SSModelDynamicGompertz class}
\usage{
\method{print}{SSModelDynamicGompertz}(model)
}
\arguments{
\item{model}{A \code{SSModelDynamicGompertz} object}
}
\description{
Accessor method to print a short description for the objects of
\code{SSModelDynamicGompertz} class
}
