% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{print.SSModelLeadingIndicator}
\alias{print.SSModelLeadingIndicator}
\title{Calling print method for classes in tsgc}
\usage{
\method{print}{SSModelLeadingIndicator}(model)
}
\arguments{
\item{model}{A \code{SSModelLeadingIndicator} object}
}
\description{
Accessor method to print a short description for the objects of
\code{SSModelLeadingIndicator} class
}
\examples{
library(tsgc)

# Specify a model
out_eng <- tsgc::SSModelLeadingIndicator(
Y = england[, 1:2], n.lag = 4, sea.period = 7,
start.date = as.Date("2021-04-30"), end.date = as.Date("2021-07-24"))

# Print a short description of the model object
print(out_eng)


}
