% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{print.SSModelLeadingIndicator}
\alias{print.SSModelLeadingIndicator}
\title{Calling print method for classes in tsgc}
\usage{
\method{print}{SSModelLeadingIndicator}(model)
}
\arguments{
\item{model}{A \code{SSModelLeadingIndicator} object}
}
\description{
Accessor method to print a short description for the objects of
\code{SSModelLeadingIndicator} class
}
