% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{Ptt}
\alias{Ptt}
\title{Extract error covariance matrix used in KFS}
\usage{
Ptt(object)
}
\arguments{
\item{object}{KFS object}
}
\description{
Accessor method to access the non-diffuse part of the error
covariance matrix in KFS object
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
idx.est <- zoo::index(gauteng) <= as.Date("2020-07-20")
# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng[idx.est], q = 0.005)
# Estimate a specified model
res <- estimate(model)

#Return filtered state estimates
Ptt(output(res))

}
