% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qtr2date}
\alias{qtr2date}
\title{Change yearqtr object into date object}
\usage{
qtr2date(dates)
}
\arguments{
\item{dates}{A vector of dates.}
}
\value{
A vector of transformed dates
}
\description{
This function changes the yearqtr or yearmon object into a date object,
e.g. the first day of the quarter/month, in order to make plotting these dates
possible in ggplot.
}
\examples{
library(xts)
my.date=yearqtr(2024.25)
qtr2date(my.date)

my.date2=yearmon(c(2024.5,2024.8))
qtr2date(my.date2)

}
