% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reinitialise_dataframe}
\alias{reinitialise_dataframe}
\title{Reinitialise a data frame by subtracting the \code{reinit.date} row from
all columns}
\usage{
reinitialise_dataframe(dt, reinit.date)
}
\arguments{
\item{dt}{Cumulated data series.}

\item{reinit.date}{Reinitialisation date. E.g. \samp{'2021-05-12'}.}
}
\value{
The reinitialised data frame
}
\description{
Reinitialise a data frame by subtracting the \code{reinit.date} row from
all columns
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
reinitialise_dataframe(gauteng,as.Date("2021-01-01"))

}
