% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/utils_old.R
\name{reinitialise_dataframe}
\alias{reinitialise_dataframe}
\title{Reinitialise a data frame by subtracting the \code{reinit.date} row from
all columns}
\usage{
reinitialise_dataframe(dt, reinit.date)

reinitialise_dataframe(dt, reinit.date)
}
\arguments{
\item{dt}{Cumulated data series, belonging to the xts class. Must only contain
1 data column in addition to a date index.}

\item{reinit.date}{Reinitialisation date, belonging to Date, yearmon or yearqtr classes. E.g. \samp{as.Date('2021-05-12')}.}
}
\value{
The reinitialised data frame

The reinitialised data frame
}
\description{
Reinitialise a data frame by subtracting the \code{reinit.date} row from
all columns

Reinitialise a data frame by subtracting the \code{reinit.date} row from
all columns
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
reinitialise_dataframe(gauteng,as.Date("2021-01-01"))

library(tsgc)
data(gauteng,package="tsgc")
reinitialise_dataframe(gauteng,as.Date("2021-01-01"))

}
