% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{seasonalComp}
\alias{seasonalComp}
\title{Extract number of seasonal components used in a KFS object}
\usage{
seasonalComp(object)
}
\arguments{
\item{object}{A \code{KFS} object.}
}
\value{
The seasonal component specification stored on the model terms.
}
\description{
Accessor method to access the number of seasonal components
used in a \code{KFS} object.
}
\examples{
library(tsgc)
data(gauteng, package = "tsgc")
model <- SSModelDynamicGompertz$new(Y = gauteng, q = 0.005,
                                    end.date = as.Date("2020-07-20"))
res <- estimate(model)
seasonalComp(output(res))

}
