% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{seasonalComp}
\alias{seasonalComp}
\title{Extract number of seasonal components used in KFS}
\usage{
seasonalComp(object)
}
\arguments{
\item{object}{KFS object}
}
\description{
Accessor method to access the number of seasonal component used in KFS object
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
idx.est <- zoo::index(gauteng) <= as.Date("2020-07-20")
# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng[idx.est], q = 0.005)
# Estimate a specified model
res <- estimate(model)

#Return number of seasonal components
seasonalComp(output(res))

}
