% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_dates}
\alias{seq_dates}
\title{Regularly spaced time objects}
\usage{
seq_dates(from, resolution, to = NA, length.out = NA)
}
\arguments{
\item{from}{The first date in the sequence.}

\item{resolution}{The time resolution of the date sequence. Options are
"daily", "monthly", "quarterly" and "yearly".}

\item{to}{The end date in the sequence, optional. If supplied, \code{to} must be after
(later than) \code{from}. Must be supplied when \code{length.out} is not provided.}

\item{length.out}{An integer for the length of the sequence, optional.
Must be supplied when \code{to} is not provided. Should not specify both
\code{length.out} and \code{to}.}
}
\value{
A vector of dates.
}
\description{
Creates a sequence of equally spaced dates that is of the class
"yearqtr", "date" or "yearmon".
}
\examples{
#Daily frequency
seq_dates(as.Date("2024-01-05"), "daily", length.out=14)

#Quarterly frequency
seq_dates(zoo::yearqtr(2020), "quarterly", length.out=12)
seq_dates(zoo::yearqtr(2020), "quarterly", to=yearqtr(2022))

#Monthly frequency
seq_dates(zoo::yearmon(2020), "monthly", length.out=12)

}
