% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSModelDynGompertz.R
\docType{class}
\name{SSModelDynamicGompertz-class}
\alias{SSModelDynamicGompertz-class}
\alias{SSModelDynamicGompertz}
\title{Class for Dynamic Gompertz Curve State-Space Model}
\description{
Class for Dynamic Gompertz Curve State-Space Model Object, which encapsulates
model settings and provides methods to obtain a FilterResults object and plot time series.

The dynamic Gompertz model with an integrated random walk (IRW) trend is defined as:
\deqn{\ln g_{t}= \delta_{t} + \varepsilon_{t}, \quad
\varepsilon_{t} \sim NID(0, \sigma_{\varepsilon}^{2}), \quad
t=2, ..., T,}
where \eqn{Y_t} is the cumulative variable, \eqn{y_t = \Delta Y_t}, and
\deqn{\ln g_{t} = \ln y_{t} - \ln Y_{t-1}.}
The trend component follows:
\deqn{\delta_{t} = \delta_{t-1} + \gamma_{t-1},}
\deqn{\gamma_{t} = \gamma_{t-1} + \zeta_{t}, \quad
\zeta_{t} \sim NID(0, \sigma_{\zeta}^{2}).}
Here, the observation disturbances \eqn{\varepsilon_{t}} and slope disturbances \eqn{\zeta_{t}} are independent and normally distributed. The signal-to-noise ratio,
\eqn{q_{\zeta} = \sigma_{\zeta}^{2} / \sigma_{\varepsilon}^{2}},
determines how rapidly the slope adjusts to new observations—higher values lead to faster changes, while lower values induce smoothness.
For models without seasonal terms (\code{sea.type = 'none'}), the priors are given by:
\deqn{\begin{pmatrix} \delta_1 \ \gamma_1 \end{pmatrix}
\sim N(a_1, P_1).}

The diffuse prior is defined as \eqn{P_1 = \kappa I_{2\times 2}} with \eqn{\kappa \to \infty}, implemented via the \code{KFAS} package (Helske, 2017). For models with a seasonal component (\code{sea.type = 'trigonometric'}), the prior mean vector \eqn{a_1} and prior covariance matrix \eqn{P_1} are extended accordingly.

See the vignette for details on the variance matrix \eqn{Q} and the observation noise variance \eqn{Q}, \eqn{H = \sigma^2_{\varepsilon}}.

This class also supports the implementation of the reinitialisation
procedure, described in the vignette and also summarised below.
Let \eqn{t=r} denote the re-initialization date and \eqn{r_0} denote the
date at which the cumulative series is set to 0. As the growth rate of
cumulative cases is defined as \eqn{g_t\equiv \frac{y_t}{Y_{t-1}}}, we have:
\deqn{\ln g_t = \ln y_t - \ln Y_{t-1} \;\;\;\; t=1, \ldots, r}
\deqn{\ln g_t^r = \ln y_t - \ln Y_{t-1}^r \;\;\;\; t=r+1, \ldots, T}
\deqn{Y_{t}^{r}=Y_{t-1}^{r}+y_{t}  \;\;\;\; t=r,\ldots,T}
where \eqn{Y_{t}^{r}} is the cumulative cases after re-initialization. We
choose to set the cumulative cases to zero at \eqn{r_0=r-1, Y_{r-1}^{r}=0},
such that the growth rate of cumulative cases is available from \eqn{t=r+1}
onwards.
We reinitialise the model by specifying the prior distribution for the
initial states appropriately. See the vignette for details.
}
\section{Fields}{

\describe{
\item{\code{Y}}{The cumulated variable.}

\item{\code{q}}{The signal-to-noise ratio (ratio of slope to irregular
variance). Defaults to \code{'NULL'}, in which case no
signal-to-noise ratio will be imposed. Instead, it will be estimated.}

\item{\code{sea.type}}{Seasonal type. Options are \code{'trigonometric'}
and \code{'none'}. \code{'trigonometric'} will yield a model with a
trigonometric seasonal component and \code{'none'} will yield a model
with no seasonal component.}

\item{\code{sea.period}}{The period of seasonality. For a day-of-the-week
effect with daily data, this would be 7. Not required if
\code{sea.type = 'none'}.}

\item{\code{reinit.date}}{(Only needed for reinitialization.) The reinitialisation date \eqn{r}. Should be
specified as an object of class \code{"Date"}. Defaults to NA, which
represents the non-reinitialized version.}

\item{\code{original.results}}{(Only needed for reinitialization.) Rather than re-estimating the model up
to the \code{reinit.date}, a \code{FilterResults} class object can be
specified here and the parameters for the reinitialisation will be taken
from this object. Default is \code{NULL}. This parameter is optional.}

\item{\code{use.presample.info}}{(Only needed for reinitialization.) Logical value denoting whether or
not to use information from before the reinitialisation date in the
reinitialisation procedure. Default is \code{TRUE}. If \code{FALSE}, the
model is estimated from scratch from the reinitialisation date and no
attempt to use information from before the reinitialisation date is made.}
}}

\section{Methods}{

\describe{
\item{\code{estimate()}}{Estimates the dynamic Gompertz curve model when applied to an object of
class \code{SSModelDynamicGompertz}.
\subsection{Return Value}{An object of class \code{FilterResults}
containing the result output for the estimated dynamic Gompertz curve
model.}}

\item{\code{initialize(
  Y,
  q = NULL,
  sea.type = "trigonometric",
  sea.period = 7,
  reinit.date = NULL,
  original.results = NULL,
  use.presample.info = TRUE
)}}{Create an instance of the \code{SSModelDynamicGompertz} class. Parameters 
are defined in `fields` section. 
  \subsection{Usage}{\code{SSModelDynGompertzReinit$new(y, q = 0.005,
  reinit.date = as.Date("2021-05-12",format = date.format))}}}

\item{\code{plot(title = NULL, series.name = "target variable", MA = TRUE)}}{Plots the lagged differences of the cumulated dataset \code{Y} in this 
      \code{SSModelDynamicGompertz} object against time, which could represent 
      daily cases.
      \subsection{Parameters}{\itemize{
        \item{\code{MA} A logical value indicating whether 7-day centered moving 
        average should be plotted. Defaults to \code{TRUE}.
        }
        \item{\code{title} Title for forecast plot. Enter as text string. 
        \code{NULL} (i.e. no title) by default.}
        \item{\code{series.name} The name of the series the growth rate is being computed
        for. E.g. \code{'cases'}. Default is `target variable`.}
}
      }
      \subsection{Return Value}{A plot of the lagged differences of the 
      cumulated dataset \code{Y} against time.}
      }

\item{\code{print()}}{Provides a quick description of SSModelDynamicGompertz object, providing 
model states and standard errors.}

\item{\code{summary()}}{Supplies details of the SSModelDynamicGompertz object, such as estimated 
parameter values, start and end dates of estimation.}
}}

\examples{
library(tsgc)
data(gauteng,package="tsgc")
idx.est <- zoo::index(gauteng) <= as.Date("2020-07-06")

# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng[idx.est], q = 0.005)

# Show summary of the model object
summary(model)

# Print a short description of the model object
print(model)

# Plot the time series in the model object
plot(model,title="Daily COVID cases in Gauteng", series.name="cases", MA=TRUE)

# Estimate a specified model
res <- estimate(model)
res

}
