% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSModelDynGompertz.R
\docType{class}
\name{SSModelDynamicGompertz-class}
\alias{SSModelDynamicGompertz-class}
\alias{SSModelDynamicGompertz}
\title{Class for Dynamic Gompertz Curve State-Space Model}
\description{
Class for Dynamic Gompertz Curve State-Space Model Object, whose
fields contain the model settings. Contains methods to obtain FilterResults
object and plot the time series.
}
\section{Fields}{

\describe{
\item{\code{Y}}{The cumulated variable.}

\item{\code{q}}{The signal-to-noise ratio (ratio of slope to irregular
variance). Defaults to \code{'NULL'}, in which case no
signal-to-noise ratio will be imposed. Instead, it will be estimated.}

\item{\code{sea.type}}{Seasonal type. Options are \code{'trigonometric'}
and \code{'none'}. \code{'trigonometric'} will yield a model with a
trigonometric seasonal component and \code{'none'} will yield a model
with no seasonal component.}

\item{\code{sea.period}}{The period of seasonality. For a day-of-the-week
effect with daily data, this would be 7. Not required if
\code{sea.type = 'none'}.}

\item{\code{reinit.date}}{(Only needed for reinitialization.) The reinitialisation date \eqn{r}. Should be
specified as an object of class \code{"Date"}. Defaults to NA, which
represents the non-reinitialized version.}

\item{\code{original.results}}{(Only needed for reinitialization.) Rather than re-estimating the model up
to the \code{reinit.date}, a \code{FilterResults} class object can be
specified here and the parameters for the reinitialisation will be taken
from this object. Default is \code{NULL}. This parameter is optional.}

\item{\code{use.presample.info}}{(Only needed for reinitialization.) Logical value denoting whether or
not to use information from before the reinitialisation date in the
reinitialisation procedure. Default is \code{TRUE}. If \code{FALSE}, the
model is estimated from scratch from the reinitialisation date and no
attempt to use information from before the reinitialisation date is made.}
}}

\section{Methods}{

\describe{
\item{\code{estimate()}}{Estimates the dynamic Gompertz curve model when applied to an object of
class \code{SSModelDynamicGompertz}.
\subsection{Return Value}{An object of class \code{FilterResults}
containing the result output for the estimated dynamic Gompertz curve
model.}}

\item{\code{get_dynamic_gompertz_model(
  y,
  q = NULL,
  sea.type = "trigonometric",
  sea.period = 7,
  a1 = NULL,
  P1 = NULL,
  Q = NULL,
  H = NULL
)}}{Returns the Dynamic Gompertz Model object, assuming no reinitialization. 
Used in the \code{get_model} function.}

\item{\code{get_model(y, q = NULL, sea.type = "trigonometric", sea.period = 7)}}{Returns dynamic Gompertz curve model. 
    \subsection{Parameters}{\itemize{
      \item{\code{y} The log-growth rate time series.}
      \item{\code{q} The signal-to-noise ratio (ratio of slope to irregular
      variance). Defaults to \code{'NULL'}, in which case no signal-to-noise
      ratio will be imposed. Instead, it will be estimated.}
      \item{\code{sea.type} Seasonal type. Options are \code{'trigonometric'}
       and \code{'none'}. \code{'trigonometric'} will yield a model with a
       trigonometric seasonal component and \code{'none'} will yield a model
       with no seasonal component.}
      \item{\code{sea.period} The period of seasonality. For a day-of-the-week
       effect with daily data, this would be 7. Not required if
       \code{sea.type = 'none'}.}
      \item{\code{a1} Optional parameter specifying the prior mean of the
      states. Defaults to \code{'NULL'}. Leave as \code{'NULL'} for a diffuse
      prior (no prior information). If a proper prior is to be specified, both
      \code{a1} and \code{P1} must be given.}
      \item{\code{P1} Optional parameter specifying the prior mean of the
      states. Defaults to \code{'NULL'}. Leave as \code{'NULL'} for a diffuse
       prior (no prior information). If a proper prior is to be specified,
       both \code{a1} and \code{P1} must be given.}
      \item{\code{Q} Optional parameter specifying the state error variances
      where these are to be imposed rather than estimated. Defaults to
      \code{'NULL'} which will see the variances estimated.}
      \item{\code{H} Optional parameter specifying the irregular variance
      where this is to be imposed rather than estimated. Defaults to
      \code{'NULL'} which will see the variance estimated.}
    }}
    \subsection{Description}{
    The dynamic Gompertz with an integrated random walk (IRW) trend is
    \deqn{\ln g_{t}=\delta_{t}+\varepsilon_{t},  \;\;\;\;
    \varepsilon_{t}\sim NID(0,\sigma_{\varepsilon }^{2}), \;\;\;\;
    t=2,...,T, }
    where \eqn{Y_t} is the cumulated variable, \eqn{y_t = \Delta Y_t},
    \eqn{\ln g_{t}=\ln y_{t}-\ln Y_{t-1}} and
    \deqn{\delta_{t} =\delta_{t-1}+\gamma_{t-1},}
    \deqn{\gamma_{t} =\gamma_{t-1}+\zeta_{t}, \;\;\;\;
    \zeta_{t}\sim NID(0,\sigma_{\zeta }^{2}),}
    where the observation disturbances \eqn{\varepsilon_{t}}  and slope
    disturbances \eqn{\zeta_{t}}, are iid Normal and mutually independent.
    Note that, the larger the signal-to-noise ratio,
    \eqn{q_{\zeta }=\sigma_{\zeta }^{2}/\sigma_{\varepsilon }^{2}},
    the faster the slope changes in response to new observations. Conversely,
    a lower signal-to-noise ratio induces smoothness.

    For the model without seasonal terms (\code{sea.type = 'none'}) the are
    priors are
    \deqn{\begin{pmatrix} \delta_1 \ \gamma_1 \end{pmatrix}
    \sim N(a_1,P_1)}.
    The diffuse prior has \eqn{P_1 = \kappa I_{2\times 2}} with
    \eqn{\kappa \to \infty}. Implementation of the diffuse prior is handled
     by the package \code{KFAS} (Helske, 2017). Where the model has a seasonal
      component (\code{sea.type = 'trigonometric'}), the vector of prior means
       \eqn{a_1} and the prior covariance matrix \eqn{P_1} are extended
       accordingly.

    See the vignette for details of the variance matrix \eqn{Q}.
    \eqn{H = \sigma^2_{\varepsilon}}.
    }
      \subsection{Return Value}{\code{KFS} model object.}}

\item{\code{initialize(
  Y,
  q = NULL,
  sea.type = "trigonometric",
  sea.period = 7,
  reinit.date = NULL,
  original.results = NULL,
  use.presample.info = TRUE
)}}{Create an instance of the \code{SSModelDynamicGompertz} class. Parameters 
are the defined in `fields` section. 
  \subsection{Usage}{\code{SSModelDynGompertzReinit$new(y, q = 0.005,
  reinit.date = as.Date("2021-05-12",format = date.format))}}}

\item{\code{plot(title = NULL, series.name = "target variable", MA = TRUE)}}{Plots the lagged differences of the cumulated dataset \code{Y} in this 
      \code{SSModelDynamicGompertz} object against time, which could represent 
      daily cases.
      \subsection{Parameters}{\itemize{
        \item{\code{MA} A logical value indicating whether 7-day centered moving 
        average should be plotted. Defaults to \code{TRUE}.
        }
        \item{\code{title} Title for forecast plot. Enter as text string. 
        \code{NULL} (i.e. no title) by default.}
        \item{\code{series.name} The name of the series the growth rate is being computed
        for. E.g. \code{'cases'}. Default is `target variable`.}
}
      }
      \subsection{Return Value}{A plot of the lagged differences of the 
      cumulated dataset \code{Y} against time.}
      }

\item{\code{print()}}{Provides a quick description of SSModelDynamicGompertz object, providing 
model states and standard errors.}

\item{\code{summary()}}{Supplies details of the SSModelDynamicGompertz object, such as estimated 
parameter values, start and end dates of estimation.}

\item{\code{update(pars, model, q, sea.type)}}{Update method for Kalman filter to implement the dynamic Gompertz curve
 model.
 A maximum of 3 parameters are used to set the observation noise
 (1 parameter), the transition equation slope and seasonal noise. If q (signal
  to noise ratio) is not null then the slope noise is set using this
  ratio.
 \subsection{Parameters}{\itemize{
  \item{\code{pars} Vector of parameters.}
  \item{\code{model} \code{KFS} model object.}
  \item{\code{q} The signal-to-noise ratio (ratio of slope to irregular
   variance).}
  \item{\code{sea.type} Seasonal type. Options are
  \code{'trigonometric'} and \code{'none'}.}
}}
\subsection{Return Value}{\code{KFS} model object.}}
}}

\examples{
library(tsgc)
data(gauteng,package="tsgc")
idx.est <- zoo::index(gauteng) <= as.Date("2020-07-06")

# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng[idx.est], q = 0.005)
# Estimate a specified model
res <- model$estimate()

}
