% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{summary.FilterResults}
\alias{summary.FilterResults}
\title{Calling summary method for FilterResults}
\usage{
\method{summary}{FilterResults}(model)
}
\arguments{
\item{model}{A \code{FilterResults} object}
}
\description{
Accessor method to show a summary for the objects of
\code{FilterResults} class
}
\examples{
library(tsgc)
data(gauteng,package="tsgc")
# Specify a model
model <- SSModelDynamicGompertz$new(Y = gauteng, q = 0.005, end.date=as.Date("2020-07-20"))
# Estimate a specified model
res <- estimate(model)

# Return KFS object in output of res
summary(res)

}
