% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{summary.FilterResults}
\alias{summary.FilterResults}
\title{Calling summary method for FilterResults}
\usage{
\method{summary}{FilterResults}(model)
}
\arguments{
\item{model}{A \code{FilterResults} object}
}
\description{
Accessor method to show a summary for the objects of
\code{FilterResults} class
}
