% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{summary.FilterResultsLI}
\alias{summary.FilterResultsLI}
\title{Calling summary method for FilterResultsLI}
\usage{
\method{summary}{FilterResultsLI}(object, ...)
}
\arguments{
\item{object}{A \code{FilterResultsLI} object}

\item{...}{Additional arguments.}
}
\description{
Accessor method to show a summary for the objects of
\code{FilterResultsLI} class
}
\examples{
library(tsgc)

out_eng <- tsgc::SSModelLeadingIndicator(
Y = england[, 1:2], n.lag = 4, sea.period = 7,
start.date = as.Date("2021-04-30"), end.date = as.Date("2021-07-24"))

res_eng<-estimate(out_eng)
summary(res_eng)

}
