% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{summary.FilterResultsLI}
\alias{summary.FilterResultsLI}
\title{Calling summary method for FilterResultsLI}
\usage{
\method{summary}{FilterResultsLI}(model)
}
\arguments{
\item{model}{A \code{FilterResultsLI} object}
}
\description{
Accessor method to show a summary for the objects of
\code{FilterResultsLI} class
}
