% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{summary.SSModelDynamicGompertz}
\alias{summary.SSModelDynamicGompertz}
\title{Calling summary method for SSModelDynamicGompertz class}
\usage{
\method{summary}{SSModelDynamicGompertz}(model)
}
\arguments{
\item{model}{A \code{SSModelDynamicGompertz} object}
}
\description{
Accessor method to show a summary for the objects of
\code{SSModelDynamicGompertz} class
}
