% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.SSModelLeadingIndicator}
\alias{summary.SSModelLeadingIndicator}
\title{Calling summary method for classes in tsgc}
\usage{
\method{summary}{SSModelLeadingIndicator}(object, ...)
}
\arguments{
\item{object}{A \code{SSModelLeadingIndicator} object}

\item{...}{Additional arguments.}
}
\description{
Accessor method to show a summary for the objects of
\code{SSModelLeadingIndicator} class
}
\examples{
library(tsgc)
data(england, package = "tsgc")
conv <- xts_to_idx(england[, 1:2])

# Specify a model
out_eng <- SSModelLeadingIndicator(
  Y = conv$series, n.lag = 4, sea.period = 7, LeadIndCol = 1,
  calendar = conv$calendar,
  start = idx_to_pos(conv$calendar, "2021-04-30"),
  end = idx_to_pos(conv$calendar, "2021-07-24"))

summary(out_eng)

}
