% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{summary.SSModelLeadingIndicator}
\alias{summary.SSModelLeadingIndicator}
\title{Calling summary method for classes in tsgc}
\usage{
\method{summary}{SSModelLeadingIndicator}(model)
}
\arguments{
\item{model}{A \code{SSModelLeadingIndicator} object}
}
\description{
Accessor method to show a summary for the objects of
\code{SSModelLeadingIndicator} class
}
