% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFns.R
\name{supply_xpred.new}
\alias{supply_xpred.new}
\title{Extract output of FilterResults}
\usage{
supply_xpred.new(object, new.xts)
}
\arguments{
\item{object}{FilterResults object}

\item{new.xts}{An xts object containing new exogenous predictors}
}
\value{
The fitted KFS model
}
\description{
Accessor method to provide new exogenous predictors for the
estimation period
}
