% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukitaly}
\alias{ukitaly}
\title{Cumulative daily cases of Covid-19 in Italy and UK, before 14 Dec 2020.}
\format{
An object of class \code{"xts"} with columns:
\describe{
\item{UK}{Cumulative cases of Covid-19 in the UK}
\item{Italy}{Cumulative cases of Covid-19 in Italy}
}
}
\usage{
data(ukitaly)
}
\description{
Cumulative daily cases of Covid-19 in Italy and UK, before 14 Dec 2020.
}
\examples{
data(ukitaly)
# plot daily cases
plot(diff(ukitaly))
}
\references{
Downloaded from https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide
}
\keyword{datasets}
