% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_results}
\alias{write_results}
\title{Write a selection of relevant results to disk}
\usage{
write_results(res, res.dir, n.ahead, prefix = "", confidence.level = 0.68)
}
\arguments{
\item{res}{Results object of class \code{FilterResults} or \code{FilterResultsLI},
obtained from \samp{estimate()} method.}

\item{res.dir}{File path to save the results to. A character string.}

\item{n.ahead}{Number of periods ahead to forecast. A positive integer.}

\item{prefix}{The prefix to be added to the file names generated. A character string.}

\item{confidence.level}{Confidence level to use for the confidence interval
on the forecasts \eqn{\ln(g_t)}.}
}
\value{
A number of csv files saved in the directory specified in
\code{res.dir}.
}
\description{
Function writes the following results to csv files which get
saved in the location specified in \code{res.dir}: forecast new cases or
incidence variable, \eqn{y}; the filtered level and slope of \eqn{\ln g},
\eqn{\delta} and \eqn{\gamma}; filtered estimates of \eqn{g_y} and the
confidence intervals for these estimates.
}
\examples{
# Not run as do not wish to save to local disk when compiling documentation.
# Below will run if copied and pasted into console.
library(tsgc)
library(here)

res.dir <- tempdir()
data(gauteng,package="tsgc")
res <- estimate(SSModelDynamicGompertz$new(Y = gauteng, q = 0.005,
end.date=as.Date("2020-07-06")))

tsgc::write_results(
res=res, res.dir = res.dir, prefix="dyn_gompertz",n.ahead = 14,
confidence.level = 0.68)

}
